Graphics.GL.Raw.Extension.OES.Texture3D

Extension Support

gl_OES_texture_3D :: Bool

Checks that the GL_OES_texture_3D extension is available.

GL_OES_texture_3D

glCompressedTexImage3DOES :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage3DOES target level internalformat width height depth border imageSize data

The length of data should be imageSize.

This command is an alias for glCompressedTexImage3D.

glCompressedTexSubImage3DOES :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage3DOES target level xoffset yoffset zoffset width height depth format imageSize data

The length of data should be imageSize.

This command is an alias for glCompressedTexSubImage3D.

glCopyTexSubImage3DOES :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage3DOES target level xoffset yoffset zoffset x y width height

This command is an alias for glCopyTexSubImage3D.

glFramebufferTexture3DOES :: MonadIO m => GLenum -> GLenum -> GLenum -> GLuint -> GLint -> GLint -> m ()

Usage: glFramebufferTexture3DOES target attachment textarget texture level zoffset

This command is an alias for glFramebufferTexture3D.

glTexImage3DOES :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexImage3DOES target level internalformat width height depth border format type pixels

The length of pixels should be COMPSIZE(format,type,width,height,depth).

This command is an alias for glTexImage3D.

glTexSubImage3DOES :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexSubImage3DOES target level xoffset yoffset zoffset width height depth format type pixels

The length of pixels should be COMPSIZE(format,type,width,height,depth).

This command is an alias for glTexSubImage3D.

pattern GL_FRAMEBUFFER_ATTACHMENT_TEXTURE_3D_ZOFFSET_OES :: (Num a, Eq a) => a

pattern GL_MAX_3D_TEXTURE_SIZE_OES :: (Num a, Eq a) => a

pattern GL_SAMPLER_3D_OES :: (Num a, Eq a) => a

pattern GL_TEXTURE_3D_OES :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_3D_OES :: (Num a, Eq a) => a

pattern GL_TEXTURE_WRAP_R_OES :: (Num a, Eq a) => a