Graphics.GL.Raw.Extension.SUN.Vertex

Extension Support

gl_SUN_vertex :: Bool

GL_SUN_vertex

glColor3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glColor3fVertex3fSUN r g b x y z

glColor3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glColor3fVertex3fvSUN c v

The length of c should be 3

The length of v should be 3

glColor4fNormal3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glColor4fNormal3fVertex3fSUN r g b a nx ny nz x y z

glColor4fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glColor4fNormal3fVertex3fvSUN c n v

The length of c should be 4

The length of n should be 3

The length of v should be 3

glColor4ubVertex2fSUN :: MonadIO m => GLubyte -> GLubyte -> GLubyte -> GLubyte -> GLfloat -> GLfloat -> m ()

Usage: glColor4ubVertex2fSUN r g b a x y

glColor4ubVertex2fvSUN :: MonadIO m => Ptr GLubyte -> Ptr GLfloat -> m ()

Usage: glColor4ubVertex2fvSUN c v

The length of c should be 4

The length of v should be 2

glColor4ubVertex3fSUN :: MonadIO m => GLubyte -> GLubyte -> GLubyte -> GLubyte -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glColor4ubVertex3fSUN r g b a x y z

glColor4ubVertex3fvSUN :: MonadIO m => Ptr GLubyte -> Ptr GLfloat -> m ()

Usage: glColor4ubVertex3fvSUN c v

The length of c should be 4

The length of v should be 3

glNormal3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glNormal3fVertex3fSUN nx ny nz x y z

glNormal3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glNormal3fVertex3fvSUN n v

The length of n should be 3

The length of v should be 3

glReplacementCodeuiColor3fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiColor3fVertex3fSUN rc r g b x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiColor3fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiColor3fVertex3fvSUN rc c v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of c should be 3

The length of v should be 3

glReplacementCodeuiColor4fNormal3fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiColor4fNormal3fVertex3fSUN rc r g b a nx ny nz x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiColor4fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiColor4fNormal3fVertex3fvSUN rc c n v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of c should be 4

The length of n should be 3

The length of v should be 3

glReplacementCodeuiColor4ubVertex3fSUN :: MonadIO m => GLuint -> GLubyte -> GLubyte -> GLubyte -> GLubyte -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiColor4ubVertex3fSUN rc r g b a x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiColor4ubVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLubyte -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiColor4ubVertex3fvSUN rc c v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of c should be 4

The length of v should be 3

glReplacementCodeuiNormal3fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiNormal3fVertex3fSUN rc nx ny nz x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiNormal3fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiNormal3fVertex3fvSUN rc n v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of n should be 3

The length of v should be 3

glReplacementCodeuiTexCoord2fColor4fNormal3fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fColor4fNormal3fVertex3fSUN rc s t r g b a nx ny nz x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiTexCoord2fColor4fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fColor4fNormal3fVertex3fvSUN rc tc c n v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of tc should be 2

The length of c should be 4

The length of n should be 3

The length of v should be 3

glReplacementCodeuiTexCoord2fNormal3fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fNormal3fVertex3fSUN rc s t nx ny nz x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiTexCoord2fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fNormal3fVertex3fvSUN rc tc n v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of tc should be 2

The length of n should be 3

The length of v should be 3

glReplacementCodeuiTexCoord2fVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fVertex3fSUN rc s t x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiTexCoord2fVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiTexCoord2fVertex3fvSUN rc tc v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of tc should be 2

The length of v should be 3

glReplacementCodeuiVertex3fSUN :: MonadIO m => GLuint -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glReplacementCodeuiVertex3fSUN rc x y z

The parameter rc is a ReplacementCodeSUN

glReplacementCodeuiVertex3fvSUN :: MonadIO m => Ptr GLuint -> Ptr GLfloat -> m ()

Usage: glReplacementCodeuiVertex3fvSUN rc v

The parameter rc is a ReplacementCodeSUN

The length of rc should be 1

The length of v should be 3

glTexCoord2fColor3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord2fColor3fVertex3fSUN s t r g b x y z

glTexCoord2fColor3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord2fColor3fVertex3fvSUN tc c v

The length of tc should be 2

The length of c should be 3

The length of v should be 3

glTexCoord2fColor4fNormal3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord2fColor4fNormal3fVertex3fSUN s t r g b a nx ny nz x y z

glTexCoord2fColor4fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord2fColor4fNormal3fVertex3fvSUN tc c n v

The length of tc should be 2

The length of c should be 4

The length of n should be 3

The length of v should be 3

glTexCoord2fColor4ubVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLubyte -> GLubyte -> GLubyte -> GLubyte -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord2fColor4ubVertex3fSUN s t r g b a x y z

glTexCoord2fColor4ubVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLubyte -> Ptr GLfloat -> m ()

Usage: glTexCoord2fColor4ubVertex3fvSUN tc c v

The length of tc should be 2

The length of c should be 4

The length of v should be 3

glTexCoord2fNormal3fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord2fNormal3fVertex3fSUN s t nx ny nz x y z

glTexCoord2fNormal3fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord2fNormal3fVertex3fvSUN tc n v

The length of tc should be 2

The length of n should be 3

The length of v should be 3

glTexCoord2fVertex3fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord2fVertex3fSUN s t x y z

glTexCoord2fVertex3fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord2fVertex3fvSUN tc v

The length of tc should be 2

The length of v should be 3

glTexCoord4fColor4fNormal3fVertex4fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord4fColor4fNormal3fVertex4fSUN s t p q r g b a nx ny nz x y z w

glTexCoord4fColor4fNormal3fVertex4fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord4fColor4fNormal3fVertex4fvSUN tc c n v

The length of tc should be 4

The length of c should be 4

The length of n should be 3

The length of v should be 4

glTexCoord4fVertex4fSUN :: MonadIO m => GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glTexCoord4fVertex4fSUN s t p q x y z w

glTexCoord4fVertex4fvSUN :: MonadIO m => Ptr GLfloat -> Ptr GLfloat -> m ()

Usage: glTexCoord4fVertex4fvSUN tc v

The length of tc should be 4

The length of v should be 4