Graphics.GL.Raw.Internal.Proc

Description

This module offers a portable way to retrieve OpenGL extension entries, providing a portability layer upon platform-specific mechanisms like glXGetProcAddress, wglGetProcAddress or NSAddressOfSymbol.

This internal module offers convenience functions and re-exports for OpenGL extension loading.

Documentation

getProcAddress :: String -> IO (FunPtr a)

Retrieve an OpenGL extension entry by name. Returns nullFunPtr when no extension entry with the given name was found.

type Invoker a = FunPtr a -> a

extensions :: Set String