Graphics.GL.Raw.Profile.Core43

Graphics.GL.Raw.Profile.Core43

module Graphics.GL.Raw.Profile.Core42

glBindVertexBuffer :: MonadIO m => GLuint -> GLuint -> GLintptr -> GLsizei -> m ()

Usage: glBindVertexBuffer bindingindex buffer offset stride

The parameter offset is a BufferOffset

glClearBufferData :: MonadIO m => GLenum -> GLenum -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glClearBufferData target internalformat format type data

The length of data should be COMPSIZE(format,type)

glClearBufferSubData :: MonadIO m => GLenum -> GLenum -> GLintptr -> GLsizeiptr -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glClearBufferSubData target internalformat offset size format type data

The parameter offset is a BufferOffset

The parameter size is a BufferSize

The length of data should be COMPSIZE(format,type)

glCopyImageSubData :: MonadIO m => GLuint -> GLenum -> GLint -> GLint -> GLint -> GLint -> GLuint -> GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> m ()

Usage: glCopyImageSubData srcName srcTarget srcLevel srcX srcY srcZ dstName dstTarget dstLevel dstX dstY dstZ srcWidth srcHeight srcDepth

glDebugMessageCallback :: MonadIO m => GLDEBUGPROC -> Ptr () -> m ()

Usage: glDebugMessageCallback callback userParam

glDebugMessageControl :: MonadIO m => GLenum -> GLenum -> GLenum -> GLsizei -> Ptr GLuint -> GLboolean -> m ()

Usage: glDebugMessageControl source type severity count ids enabled

The parameter enabled is a Boolean, one of: GL_FALSE, GL_TRUE

The length of ids should be count

glDebugMessageInsert :: MonadIO m => GLenum -> GLenum -> GLuint -> GLenum -> GLsizei -> Ptr GLchar -> m ()

Usage: glDebugMessageInsert source type id severity length buf

The length of buf should be COMPSIZE(buf,length)

glDispatchCompute :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glDispatchCompute num_groups_x num_groups_y num_groups_z

glDispatchComputeIndirect :: MonadIO m => GLintptr -> m ()

Usage: glDispatchComputeIndirect indirect

The parameter indirect is a BufferOffset

glFramebufferParameteri :: MonadIO m => GLenum -> GLenum -> GLint -> m ()

Usage: glFramebufferParameteri target pname param

glGetDebugMessageLog :: MonadIO m => GLuint -> GLsizei -> Ptr GLenum -> Ptr GLenum -> Ptr GLuint -> Ptr GLenum -> Ptr GLsizei -> Ptr GLchar -> m GLuint

Usage: glGetDebugMessageLog count bufSize sources types ids severities lengths messageLog

The length of sources should be count

The length of types should be count

The length of ids should be count

The length of severities should be count

The length of lengths should be count

The length of messageLog should be bufSize

glGetFramebufferParameteriv :: MonadIO m => GLenum -> GLenum -> Ptr GLint -> m ()

Usage: glGetFramebufferParameteriv target pname params

The length of params should be COMPSIZE(pname)

glGetInternalformati64v :: MonadIO m => GLenum -> GLenum -> GLenum -> GLsizei -> Ptr GLint64 -> m ()

Usage: glGetInternalformati64v target internalformat pname bufSize params

The length of params should be bufSize

glGetObjectLabel :: MonadIO m => GLenum -> GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> m ()

Usage: glGetObjectLabel identifier name bufSize length label

The length of length should be 1

The length of label should be bufSize

glGetObjectPtrLabel :: MonadIO m => Ptr () -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> m ()

Usage: glGetObjectPtrLabel ptr bufSize length label

The length of length should be 1

The length of label should be bufSize

glGetPointerv :: MonadIO m => GLenum -> Ptr () -> m ()

Usage: glGetPointerv pname params

The parameter pname is a GetPointervPName, one of: GL_COLOR_ARRAY_POINTER, GL_COLOR_ARRAY_POINTER_EXT, GL_EDGE_FLAG_ARRAY_POINTER, GL_EDGE_FLAG_ARRAY_POINTER_EXT, GL_FEEDBACK_BUFFER_POINTER, GL_INDEX_ARRAY_POINTER, GL_INDEX_ARRAY_POINTER_EXT, GL_INSTRUMENT_BUFFER_POINTER_SGIX, GL_NORMAL_ARRAY_POINTER, GL_NORMAL_ARRAY_POINTER_EXT, GL_SELECTION_BUFFER_POINTER, GL_TEXTURE_COORD_ARRAY_POINTER, GL_TEXTURE_COORD_ARRAY_POINTER_EXT, GL_VERTEX_ARRAY_POINTER, GL_VERTEX_ARRAY_POINTER_EXT

The length of params should be 1

glGetProgramInterfaceiv :: MonadIO m => GLuint -> GLenum -> GLenum -> Ptr GLint -> m ()

Usage: glGetProgramInterfaceiv program programInterface pname params

The length of params should be COMPSIZE(pname)

glGetProgramResourceIndex :: MonadIO m => GLuint -> GLenum -> Ptr GLchar -> m GLuint

Usage: glGetProgramResourceIndex program programInterface name

The length of name should be COMPSIZE(name)

glGetProgramResourceLocation :: MonadIO m => GLuint -> GLenum -> Ptr GLchar -> m GLint

Usage: glGetProgramResourceLocation program programInterface name

The length of name should be COMPSIZE(name)

glGetProgramResourceLocationIndex :: MonadIO m => GLuint -> GLenum -> Ptr GLchar -> m GLint

Usage: glGetProgramResourceLocationIndex program programInterface name

The length of name should be COMPSIZE(name)

glGetProgramResourceName :: MonadIO m => GLuint -> GLenum -> GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> m ()

Usage: glGetProgramResourceName program programInterface index bufSize length name

The length of length should be 1

The length of name should be bufSize

glGetProgramResourceiv :: MonadIO m => GLuint -> GLenum -> GLuint -> GLsizei -> Ptr GLenum -> GLsizei -> Ptr GLsizei -> Ptr GLint -> m ()

Usage: glGetProgramResourceiv program programInterface index propCount props bufSize length params

The length of props should be propCount

The length of length should be 1

The length of params should be bufSize

glInvalidateBufferData :: MonadIO m => GLuint -> m ()

Usage: glInvalidateBufferData buffer

glInvalidateBufferSubData :: MonadIO m => GLuint -> GLintptr -> GLsizeiptr -> m ()

Usage: glInvalidateBufferSubData buffer offset length

The parameter offset is a BufferOffset

The parameter length is a BufferSize

glInvalidateFramebuffer :: MonadIO m => GLenum -> GLsizei -> Ptr GLenum -> m ()

Usage: glInvalidateFramebuffer target numAttachments attachments

The length of attachments should be numAttachments

glInvalidateSubFramebuffer :: MonadIO m => GLenum -> GLsizei -> Ptr GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glInvalidateSubFramebuffer target numAttachments attachments x y width height

The length of attachments should be numAttachments

glInvalidateTexImage :: MonadIO m => GLuint -> GLint -> m ()

Usage: glInvalidateTexImage texture level

glInvalidateTexSubImage :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> m ()

Usage: glInvalidateTexSubImage texture level xoffset yoffset zoffset width height depth

glMultiDrawArraysIndirect :: MonadIO m => GLenum -> Ptr () -> GLsizei -> GLsizei -> m ()

Usage: glMultiDrawArraysIndirect mode indirect drawcount stride

The length of indirect should be COMPSIZE(drawcount,stride)

glMultiDrawElementsIndirect :: MonadIO m => GLenum -> GLenum -> Ptr () -> GLsizei -> GLsizei -> m ()

Usage: glMultiDrawElementsIndirect mode type indirect drawcount stride

The length of indirect should be COMPSIZE(drawcount,stride)

glObjectLabel :: MonadIO m => GLenum -> GLuint -> GLsizei -> Ptr GLchar -> m ()

Usage: glObjectLabel identifier name length label

The length of label should be COMPSIZE(label,length)

glObjectPtrLabel :: MonadIO m => Ptr () -> GLsizei -> Ptr GLchar -> m ()

Usage: glObjectPtrLabel ptr length label

The length of label should be COMPSIZE(label,length)

glPopDebugGroup :: MonadIO m => m ()

Usage: glPopDebugGroup

glPushDebugGroup :: MonadIO m => GLenum -> GLuint -> GLsizei -> Ptr GLchar -> m ()

Usage: glPushDebugGroup source id length message

The length of message should be COMPSIZE(message,length)

glShaderStorageBlockBinding :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glShaderStorageBlockBinding program storageBlockIndex storageBlockBinding

glTexBufferRange :: MonadIO m => GLenum -> GLenum -> GLuint -> GLintptr -> GLsizeiptr -> m ()

Usage: glTexBufferRange target internalformat buffer offset size

The parameter offset is a BufferOffset

The parameter size is a BufferSize

glTexStorage2DMultisample :: MonadIO m => GLenum -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLboolean -> m ()

Usage: glTexStorage2DMultisample target samples internalformat width height fixedsamplelocations

The parameter fixedsamplelocations is a Boolean, one of: GL_FALSE, GL_TRUE

glTexStorage3DMultisample :: MonadIO m => GLenum -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLboolean -> m ()

Usage: glTexStorage3DMultisample target samples internalformat width height depth fixedsamplelocations

The parameter fixedsamplelocations is a Boolean, one of: GL_FALSE, GL_TRUE

glTextureView :: MonadIO m => GLuint -> GLenum -> GLuint -> GLenum -> GLuint -> GLuint -> GLuint -> GLuint -> m ()

Usage: glTextureView texture target origtexture internalformat minlevel numlevels minlayer numlayers

glVertexAttribBinding :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glVertexAttribBinding attribindex bindingindex

glVertexAttribFormat :: MonadIO m => GLuint -> GLint -> GLenum -> GLboolean -> GLuint -> m ()

Usage: glVertexAttribFormat attribindex size type normalized relativeoffset

The parameter normalized is a Boolean, one of: GL_FALSE, GL_TRUE

glVertexAttribIFormat :: MonadIO m => GLuint -> GLint -> GLenum -> GLuint -> m ()

Usage: glVertexAttribIFormat attribindex size type relativeoffset

glVertexAttribLFormat :: MonadIO m => GLuint -> GLint -> GLenum -> GLuint -> m ()

Usage: glVertexAttribLFormat attribindex size type relativeoffset

glVertexBindingDivisor :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glVertexBindingDivisor bindingindex divisor

pattern GL_ACTIVE_RESOURCES :: (Num a, Eq a) => a

pattern GL_ACTIVE_VARIABLES :: (Num a, Eq a) => a

pattern GL_ANY_SAMPLES_PASSED_CONSERVATIVE :: (Num a, Eq a) => a

pattern GL_ARRAY_SIZE :: (Num a, Eq a) => a

pattern GL_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_ATOMIC_COUNTER_BUFFER_INDEX :: (Num a, Eq a) => a

pattern GL_ATOMIC_COUNTER_BUFFER_REFERENCED_BY_COMPUTE_SHADER :: (Num a, Eq a) => a

pattern GL_AUTO_GENERATE_MIPMAP :: (Num a, Eq a) => a

pattern GL_BLOCK_INDEX :: (Num a, Eq a) => a

pattern GL_BUFFER :: (Num a, Eq a) => a

pattern GL_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_BUFFER_DATA_SIZE :: (Num a, Eq a) => a

pattern GL_BUFFER_VARIABLE :: (Num a, Eq a) => a

pattern GL_CAVEAT_SUPPORT :: (Num a, Eq a) => a

pattern GL_CLEAR_BUFFER :: (Num a, Eq a) => a

pattern GL_COLOR_COMPONENTS :: (Num a, Eq a) => a

pattern GL_COLOR_ENCODING :: (Num a, Eq a) => a

pattern GL_COLOR_RENDERABLE :: (Num a, Eq a) => a

pattern GL_COMPRESSED_R11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RG11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB8_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB8_PUNCHTHROUGH_ALPHA1_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGBA8_ETC2_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SIGNED_R11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SIGNED_RG11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_ALPHA8_ETC2_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_PUNCHTHROUGH_ALPHA1_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPUTE_SHADER :: (Num a, Eq a) => a

pattern GL_COMPUTE_SHADER_BIT :: (Num a, Eq a) => a

pattern GL_COMPUTE_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_COMPUTE_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_COMPUTE_TEXTURE :: (Num a, Eq a) => a

pattern GL_COMPUTE_WORK_GROUP_SIZE :: (Num a, Eq a) => a

pattern GL_CONTEXT_FLAG_DEBUG_BIT :: (Num a, Eq a) => a

pattern GL_DEBUG_CALLBACK_FUNCTION :: (Num a, Eq a) => a

pattern GL_DEBUG_CALLBACK_USER_PARAM :: (Num a, Eq a) => a

pattern GL_DEBUG_GROUP_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_DEBUG_LOGGED_MESSAGES :: (Num a, Eq a) => a

pattern GL_DEBUG_NEXT_LOGGED_MESSAGE_LENGTH :: (Num a, Eq a) => a

pattern GL_DEBUG_OUTPUT :: (Num a, Eq a) => a

pattern GL_DEBUG_OUTPUT_SYNCHRONOUS :: (Num a, Eq a) => a

pattern GL_DEBUG_SEVERITY_HIGH :: (Num a, Eq a) => a

pattern GL_DEBUG_SEVERITY_LOW :: (Num a, Eq a) => a

pattern GL_DEBUG_SEVERITY_MEDIUM :: (Num a, Eq a) => a

pattern GL_DEBUG_SEVERITY_NOTIFICATION :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_API :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_APPLICATION :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_OTHER :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_SHADER_COMPILER :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_THIRD_PARTY :: (Num a, Eq a) => a

pattern GL_DEBUG_SOURCE_WINDOW_SYSTEM :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_ERROR :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_MARKER :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_OTHER :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_PERFORMANCE :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_POP_GROUP :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_PORTABILITY :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_PUSH_GROUP :: (Num a, Eq a) => a

pattern GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR :: (Num a, Eq a) => a

pattern GL_DEPTH_COMPONENTS :: (Num a, Eq a) => a

pattern GL_DEPTH_RENDERABLE :: (Num a, Eq a) => a

pattern GL_DEPTH_STENCIL_TEXTURE_MODE :: (Num a, Eq a) => a

pattern GL_DISPATCH_INDIRECT_BUFFER :: (Num a, Eq a) => a

pattern GL_DISPATCH_INDIRECT_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_FILTER :: (Num a, Eq a) => a

pattern GL_FRAGMENT_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_FRAGMENT_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_FRAGMENT_TEXTURE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_BLEND :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT_FIXED_SAMPLE_LOCATIONS :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT_HEIGHT :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT_LAYERS :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT_SAMPLES :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT_WIDTH :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_RENDERABLE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_RENDERABLE_LAYERED :: (Num a, Eq a) => a

pattern GL_FULL_SUPPORT :: (Num a, Eq a) => a

pattern GL_GEOMETRY_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_GEOMETRY_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_GEOMETRY_TEXTURE :: (Num a, Eq a) => a

pattern GL_GET_TEXTURE_IMAGE_FORMAT :: (Num a, Eq a) => a

pattern GL_GET_TEXTURE_IMAGE_TYPE :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_10_10_10_2 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_11_11_10 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_1_X_16 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_1_X_32 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_1_X_8 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_2_X_16 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_2_X_32 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_2_X_8 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_4_X_16 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_4_X_32 :: (Num a, Eq a) => a

pattern GL_IMAGE_CLASS_4_X_8 :: (Num a, Eq a) => a

pattern GL_IMAGE_COMPATIBILITY_CLASS :: (Num a, Eq a) => a

pattern GL_IMAGE_PIXEL_FORMAT :: (Num a, Eq a) => a

pattern GL_IMAGE_PIXEL_TYPE :: (Num a, Eq a) => a

pattern GL_IMAGE_TEXEL_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_ALPHA_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_ALPHA_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_BLUE_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_BLUE_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_DEPTH_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_DEPTH_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_GREEN_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_GREEN_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_PREFERRED :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_RED_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_RED_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_SHARED_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_STENCIL_SIZE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_STENCIL_TYPE :: (Num a, Eq a) => a

pattern GL_INTERNALFORMAT_SUPPORTED :: (Num a, Eq a) => a

pattern GL_IS_PER_PATCH :: (Num a, Eq a) => a

pattern GL_IS_ROW_MAJOR :: (Num a, Eq a) => a

pattern GL_LOCATION :: (Num a, Eq a) => a

pattern GL_LOCATION_INDEX :: (Num a, Eq a) => a

pattern GL_MANUAL_GENERATE_MIPMAP :: (Num a, Eq a) => a

pattern GL_MATRIX_STRIDE :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_COMPUTE_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_DIMENSIONS :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_SHADER_OUTPUT_RESOURCES :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_ATOMIC_COUNTERS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_ATOMIC_COUNTER_BUFFERS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_IMAGE_UNIFORMS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_SHARED_MEMORY_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_TEXTURE_IMAGE_UNITS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_UNIFORM_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_WORK_GROUP_COUNT :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_WORK_GROUP_INVOCATIONS :: (Num a, Eq a) => a

pattern GL_MAX_COMPUTE_WORK_GROUP_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_DEBUG_GROUP_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_DEBUG_LOGGED_MESSAGES :: (Num a, Eq a) => a

pattern GL_MAX_DEBUG_MESSAGE_LENGTH :: (Num a, Eq a) => a

pattern GL_MAX_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENT_INDEX :: (Num a, Eq a) => a

pattern GL_MAX_FRAGMENT_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_FRAMEBUFFER_HEIGHT :: (Num a, Eq a) => a

pattern GL_MAX_FRAMEBUFFER_LAYERS :: (Num a, Eq a) => a

pattern GL_MAX_FRAMEBUFFER_SAMPLES :: (Num a, Eq a) => a

pattern GL_MAX_FRAMEBUFFER_WIDTH :: (Num a, Eq a) => a

pattern GL_MAX_GEOMETRY_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_HEIGHT :: (Num a, Eq a) => a

pattern GL_MAX_LABEL_LENGTH :: (Num a, Eq a) => a

pattern GL_MAX_LAYERS :: (Num a, Eq a) => a

pattern GL_MAX_NAME_LENGTH :: (Num a, Eq a) => a

pattern GL_MAX_NUM_ACTIVE_VARIABLES :: (Num a, Eq a) => a

pattern GL_MAX_NUM_COMPATIBLE_SUBROUTINES :: (Num a, Eq a) => a

pattern GL_MAX_SHADER_STORAGE_BLOCK_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_SHADER_STORAGE_BUFFER_BINDINGS :: (Num a, Eq a) => a

pattern GL_MAX_TESS_CONTROL_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_TESS_EVALUATION_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_UNIFORM_LOCATIONS :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_ATTRIB_BINDINGS :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_ATTRIB_RELATIVE_OFFSET :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_SHADER_STORAGE_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_WIDTH :: (Num a, Eq a) => a

pattern GL_MIPMAP :: (Num a, Eq a) => a

pattern GL_NAME_LENGTH :: (Num a, Eq a) => a

pattern GL_NUM_ACTIVE_VARIABLES :: (Num a, Eq a) => a

pattern GL_NUM_SHADING_LANGUAGE_VERSIONS :: (Num a, Eq a) => a

pattern GL_OFFSET :: (Num a, Eq a) => a

pattern GL_PRIMITIVE_RESTART_FIXED_INDEX :: (Num a, Eq a) => a

pattern GL_PROGRAM :: (Num a, Eq a) => a

pattern GL_PROGRAM_INPUT :: (Num a, Eq a) => a

pattern GL_PROGRAM_OUTPUT :: (Num a, Eq a) => a

pattern GL_PROGRAM_PIPELINE :: (Num a, Eq a) => a

pattern GL_QUERY :: (Num a, Eq a) => a

pattern GL_READ_PIXELS :: (Num a, Eq a) => a

pattern GL_READ_PIXELS_FORMAT :: (Num a, Eq a) => a

pattern GL_READ_PIXELS_TYPE :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_COMPUTE_SHADER :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_FRAGMENT_SHADER :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_GEOMETRY_SHADER :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_TESS_CONTROL_SHADER :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_TESS_EVALUATION_SHADER :: (Num a, Eq a) => a

pattern GL_REFERENCED_BY_VERTEX_SHADER :: (Num a, Eq a) => a

pattern GL_SAMPLER :: (Num a, Eq a) => a

pattern GL_SHADER :: (Num a, Eq a) => a

pattern GL_SHADER_IMAGE_ATOMIC :: (Num a, Eq a) => a

pattern GL_SHADER_IMAGE_LOAD :: (Num a, Eq a) => a

pattern GL_SHADER_IMAGE_STORE :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BARRIER_BIT :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BLOCK :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BUFFER :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BUFFER_OFFSET_ALIGNMENT :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_SHADER_STORAGE_BUFFER_START :: (Num a, Eq a) => a

pattern GL_SIMULTANEOUS_TEXTURE_AND_DEPTH_TEST :: (Num a, Eq a) => a

pattern GL_SIMULTANEOUS_TEXTURE_AND_DEPTH_WRITE :: (Num a, Eq a) => a

pattern GL_SIMULTANEOUS_TEXTURE_AND_STENCIL_TEST :: (Num a, Eq a) => a

pattern GL_SIMULTANEOUS_TEXTURE_AND_STENCIL_WRITE :: (Num a, Eq a) => a

pattern GL_SRGB_READ :: (Num a, Eq a) => a

pattern GL_SRGB_WRITE :: (Num a, Eq a) => a

pattern GL_STACK_OVERFLOW :: (Num a, Eq a) => a

pattern GL_STACK_UNDERFLOW :: (Num a, Eq a) => a

pattern GL_STENCIL_COMPONENTS :: (Num a, Eq a) => a

pattern GL_STENCIL_RENDERABLE :: (Num a, Eq a) => a

pattern GL_TESS_CONTROL_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_TESS_CONTROL_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_TESS_CONTROL_TEXTURE :: (Num a, Eq a) => a

pattern GL_TESS_EVALUATION_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_TESS_EVALUATION_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_TESS_EVALUATION_TEXTURE :: (Num a, Eq a) => a

pattern GL_TEXTURE_BUFFER_OFFSET :: (Num a, Eq a) => a

pattern GL_TEXTURE_BUFFER_OFFSET_ALIGNMENT :: (Num a, Eq a) => a

pattern GL_TEXTURE_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_BLOCK_HEIGHT :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_BLOCK_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_BLOCK_WIDTH :: (Num a, Eq a) => a

pattern GL_TEXTURE_GATHER :: (Num a, Eq a) => a

pattern GL_TEXTURE_GATHER_SHADOW :: (Num a, Eq a) => a

pattern GL_TEXTURE_IMAGE_FORMAT :: (Num a, Eq a) => a

pattern GL_TEXTURE_IMAGE_TYPE :: (Num a, Eq a) => a

pattern GL_TEXTURE_IMMUTABLE_LEVELS :: (Num a, Eq a) => a

pattern GL_TEXTURE_SHADOW :: (Num a, Eq a) => a

pattern GL_TEXTURE_VIEW :: (Num a, Eq a) => a

pattern GL_TEXTURE_VIEW_MIN_LAYER :: (Num a, Eq a) => a

pattern GL_TEXTURE_VIEW_MIN_LEVEL :: (Num a, Eq a) => a

pattern GL_TEXTURE_VIEW_NUM_LAYERS :: (Num a, Eq a) => a

pattern GL_TEXTURE_VIEW_NUM_LEVELS :: (Num a, Eq a) => a

pattern GL_TOP_LEVEL_ARRAY_SIZE :: (Num a, Eq a) => a

pattern GL_TOP_LEVEL_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_VARYING :: (Num a, Eq a) => a

pattern GL_TYPE :: (Num a, Eq a) => a

pattern GL_UNIFORM :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_REFERENCED_BY_COMPUTE_SHADER :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY :: (Num a, Eq a) => a

pattern GL_VERTEX_ATTRIB_ARRAY_LONG :: (Num a, Eq a) => a

pattern GL_VERTEX_ATTRIB_BINDING :: (Num a, Eq a) => a

pattern GL_VERTEX_ATTRIB_RELATIVE_OFFSET :: (Num a, Eq a) => a

pattern GL_VERTEX_BINDING_BUFFER :: (Num a, Eq a) => a

pattern GL_VERTEX_BINDING_DIVISOR :: (Num a, Eq a) => a

pattern GL_VERTEX_BINDING_OFFSET :: (Num a, Eq a) => a

pattern GL_VERTEX_BINDING_STRIDE :: (Num a, Eq a) => a

pattern GL_VERTEX_SUBROUTINE :: (Num a, Eq a) => a

pattern GL_VERTEX_SUBROUTINE_UNIFORM :: (Num a, Eq a) => a

pattern GL_VERTEX_TEXTURE :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_128_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_16_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_24_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_32_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_48_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_64_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_8_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_96_BITS :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_BPTC_FLOAT :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_BPTC_UNORM :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_RGTC1_RED :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_RGTC2_RG :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_S3TC_DXT1_RGB :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_S3TC_DXT1_RGBA :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_S3TC_DXT3_RGBA :: (Num a, Eq a) => a

pattern GL_VIEW_CLASS_S3TC_DXT5_RGBA :: (Num a, Eq a) => a

pattern GL_VIEW_COMPATIBILITY_CLASS :: (Num a, Eq a) => a