Graphics.GL.Raw.Profile.Core45

Graphics.GL.Raw.Profile.Core45

module Graphics.GL.Raw.Profile.Core44

glBindTextureUnit :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glBindTextureUnit unit texture

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindTextureUnit.xhtml

glBlitNamedFramebuffer :: MonadIO m => GLuint -> GLuint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLbitfield -> GLenum -> m ()

Usage: glBlitNamedFramebuffer readFramebuffer drawFramebuffer srcX0 srcY0 srcX1 srcY1 dstX0 dstY0 dstX1 dstY1 mask filter

glCheckNamedFramebufferStatus :: MonadIO m => GLuint -> GLenum -> m GLenum

Usage: glCheckNamedFramebufferStatus framebuffer target

glClearNamedBufferData :: MonadIO m => GLuint -> GLenum -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glClearNamedBufferData buffer internalformat format type data

glClearNamedBufferSubData :: MonadIO m => GLuint -> GLenum -> GLintptr -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glClearNamedBufferSubData buffer internalformat offset size format type data

glClearNamedFramebufferfi :: MonadIO m => GLuint -> GLenum -> GLfloat -> GLint -> m ()

Usage: glClearNamedFramebufferfi framebuffer buffer depth stencil

glClearNamedFramebufferfv :: MonadIO m => GLuint -> GLenum -> GLint -> Ptr GLfloat -> m ()

Usage: glClearNamedFramebufferfv framebuffer buffer drawbuffer value

glClearNamedFramebufferiv :: MonadIO m => GLuint -> GLenum -> GLint -> Ptr GLint -> m ()

Usage: glClearNamedFramebufferiv framebuffer buffer drawbuffer value

glClearNamedFramebufferuiv :: MonadIO m => GLuint -> GLenum -> GLint -> Ptr GLuint -> m ()

Usage: glClearNamedFramebufferuiv framebuffer buffer drawbuffer value

glClipControl :: MonadIO m => GLenum -> GLenum -> m ()

Usage: glClipControl origin depth

Manual page: https://www.opengl.org/sdk/docs/man/html/glClipControl.xhtml

glCompressedTextureSubImage1D :: MonadIO m => GLuint -> GLint -> GLint -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTextureSubImage1D texture level xoffset width format imageSize data

glCompressedTextureSubImage2D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTextureSubImage2D texture level xoffset yoffset width height format imageSize data

glCompressedTextureSubImage3D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTextureSubImage3D texture level xoffset yoffset zoffset width height depth format imageSize data

glCopyNamedBufferSubData :: MonadIO m => GLuint -> GLuint -> GLintptr -> GLintptr -> GLsizei -> m ()

Usage: glCopyNamedBufferSubData readBuffer writeBuffer readOffset writeOffset size

glCopyTextureSubImage1D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> m ()

Usage: glCopyTextureSubImage1D texture level xoffset x y width

glCopyTextureSubImage2D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTextureSubImage2D texture level xoffset yoffset x y width height

glCopyTextureSubImage3D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTextureSubImage3D texture level xoffset yoffset zoffset x y width height

glCreateBuffers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateBuffers n buffers

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateBuffers.xhtml

glCreateFramebuffers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateFramebuffers n framebuffers

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateFramebuffers.xhtml

glCreateProgramPipelines :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateProgramPipelines n pipelines

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateProgramPipelines.xhtml

glCreateQueries :: MonadIO m => GLenum -> GLsizei -> Ptr GLuint -> m ()

Usage: glCreateQueries target n ids

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateQueries.xhtml

glCreateRenderbuffers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateRenderbuffers n renderbuffers

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateRenderbuffers.xhtml

glCreateSamplers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateSamplers n samplers

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateSamplers.xhtml

glCreateTextures :: MonadIO m => GLenum -> GLsizei -> Ptr GLuint -> m ()

Usage: glCreateTextures target n textures

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateTextures.xhtml

glCreateTransformFeedbacks :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateTransformFeedbacks n ids

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateTransformFeedbacks.xhtml

glCreateVertexArrays :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glCreateVertexArrays n arrays

Manual page: https://www.opengl.org/sdk/docs/man/html/glCreateVertexArrays.xhtml

glDisableVertexArrayAttrib :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glDisableVertexArrayAttrib vaobj index

glEnableVertexArrayAttrib :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glEnableVertexArrayAttrib vaobj index

glFlushMappedNamedBufferRange :: MonadIO m => GLuint -> GLintptr -> GLsizei -> m ()

Usage: glFlushMappedNamedBufferRange buffer offset length

glGenerateTextureMipmap :: MonadIO m => GLuint -> m ()

Usage: glGenerateTextureMipmap texture

glGetCompressedTextureImage :: MonadIO m => GLuint -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glGetCompressedTextureImage texture level bufSize pixels

glGetCompressedTextureSubImage :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLsizei -> Ptr () -> m ()

Usage: glGetCompressedTextureSubImage texture level xoffset yoffset zoffset width height depth bufSize pixels

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetCompressedTextureSubImage.xhtml

glGetGraphicsResetStatus :: MonadIO m => m GLenum

Usage: glGetGraphicsResetStatus

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetGraphicsResetStatus.xhtml

glGetNamedBufferParameteri64v :: MonadIO m => GLuint -> GLenum -> Ptr GLint64 -> m ()

Usage: glGetNamedBufferParameteri64v buffer pname params

glGetNamedBufferParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetNamedBufferParameteriv buffer pname params

glGetNamedBufferPointerv :: MonadIO m => GLuint -> GLenum -> Ptr () -> m ()

Usage: glGetNamedBufferPointerv buffer pname params

glGetNamedBufferSubData :: MonadIO m => GLuint -> GLintptr -> GLsizei -> Ptr () -> m ()

Usage: glGetNamedBufferSubData buffer offset size data

glGetNamedFramebufferAttachmentParameteriv :: MonadIO m => GLuint -> GLenum -> GLenum -> Ptr GLint -> m ()

Usage: glGetNamedFramebufferAttachmentParameteriv framebuffer attachment pname params

glGetNamedFramebufferParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetNamedFramebufferParameteriv framebuffer pname param

glGetNamedRenderbufferParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetNamedRenderbufferParameteriv renderbuffer pname params

glGetQueryBufferObjecti64v :: MonadIO m => GLuint -> GLuint -> GLenum -> GLintptr -> m ()

Usage: glGetQueryBufferObjecti64v id buffer pname offset

glGetQueryBufferObjectiv :: MonadIO m => GLuint -> GLuint -> GLenum -> GLintptr -> m ()

Usage: glGetQueryBufferObjectiv id buffer pname offset

glGetQueryBufferObjectui64v :: MonadIO m => GLuint -> GLuint -> GLenum -> GLintptr -> m ()

Usage: glGetQueryBufferObjectui64v id buffer pname offset

glGetQueryBufferObjectuiv :: MonadIO m => GLuint -> GLuint -> GLenum -> GLintptr -> m ()

Usage: glGetQueryBufferObjectuiv id buffer pname offset

glGetTextureImage :: MonadIO m => GLuint -> GLint -> GLenum -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glGetTextureImage texture level format type bufSize pixels

glGetTextureLevelParameterfv :: MonadIO m => GLuint -> GLint -> GLenum -> Ptr GLfloat -> m ()

Usage: glGetTextureLevelParameterfv texture level pname params

glGetTextureLevelParameteriv :: MonadIO m => GLuint -> GLint -> GLenum -> Ptr GLint -> m ()

Usage: glGetTextureLevelParameteriv texture level pname params

glGetTextureParameterIiv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetTextureParameterIiv texture pname params

glGetTextureParameterIuiv :: MonadIO m => GLuint -> GLenum -> Ptr GLuint -> m ()

Usage: glGetTextureParameterIuiv texture pname params

glGetTextureParameterfv :: MonadIO m => GLuint -> GLenum -> Ptr GLfloat -> m ()

Usage: glGetTextureParameterfv texture pname params

glGetTextureParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetTextureParameteriv texture pname params

glGetTextureSubImage :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glGetTextureSubImage texture level xoffset yoffset zoffset width height depth format type bufSize pixels

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetTextureSubImage.xhtml

glGetTransformFeedbacki64_v :: MonadIO m => GLuint -> GLenum -> GLuint -> Ptr GLint64 -> m ()

Usage: glGetTransformFeedbacki64_v xfb pname index param

glGetTransformFeedbacki_v :: MonadIO m => GLuint -> GLenum -> GLuint -> Ptr GLint -> m ()

Usage: glGetTransformFeedbacki_v xfb pname index param

glGetTransformFeedbackiv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetTransformFeedbackiv xfb pname param

glGetVertexArrayIndexed64iv :: MonadIO m => GLuint -> GLuint -> GLenum -> Ptr GLint64 -> m ()

Usage: glGetVertexArrayIndexed64iv vaobj index pname param

glGetVertexArrayIndexediv :: MonadIO m => GLuint -> GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetVertexArrayIndexediv vaobj index pname param

glGetVertexArrayiv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetVertexArrayiv vaobj pname param

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetVertexArrayiv.xhtml

glGetnCompressedTexImage :: MonadIO m => GLenum -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glGetnCompressedTexImage target lod bufSize pixels

glGetnTexImage :: MonadIO m => GLenum -> GLint -> GLenum -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glGetnTexImage target level format type bufSize pixels

glGetnUniformdv :: MonadIO m => GLuint -> GLint -> GLsizei -> Ptr GLdouble -> m ()

Usage: glGetnUniformdv program location bufSize params

glGetnUniformfv :: MonadIO m => GLuint -> GLint -> GLsizei -> Ptr GLfloat -> m ()

Usage: glGetnUniformfv program location bufSize params

glGetnUniformiv :: MonadIO m => GLuint -> GLint -> GLsizei -> Ptr GLint -> m ()

Usage: glGetnUniformiv program location bufSize params

glGetnUniformuiv :: MonadIO m => GLuint -> GLint -> GLsizei -> Ptr GLuint -> m ()

Usage: glGetnUniformuiv program location bufSize params

glInvalidateNamedFramebufferData :: MonadIO m => GLuint -> GLsizei -> Ptr GLenum -> m ()

Usage: glInvalidateNamedFramebufferData framebuffer numAttachments attachments

glInvalidateNamedFramebufferSubData :: MonadIO m => GLuint -> GLsizei -> Ptr GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glInvalidateNamedFramebufferSubData framebuffer numAttachments attachments x y width height

glMapNamedBuffer :: MonadIO m => GLuint -> GLenum -> m (Ptr ())

Usage: glMapNamedBuffer buffer access

glMapNamedBufferRange :: MonadIO m => GLuint -> GLintptr -> GLsizei -> GLbitfield -> m (Ptr ())

Usage: glMapNamedBufferRange buffer offset length access

glMemoryBarrierByRegion :: MonadIO m => GLbitfield -> m ()

Usage: glMemoryBarrierByRegion barriers

glNamedBufferData :: MonadIO m => GLuint -> GLsizei -> Ptr () -> GLenum -> m ()

Usage: glNamedBufferData buffer size data usage

glNamedBufferStorage :: MonadIO m => GLuint -> GLsizei -> Ptr () -> GLbitfield -> m ()

Usage: glNamedBufferStorage buffer size data flags

glNamedBufferSubData :: MonadIO m => GLuint -> GLintptr -> GLsizei -> Ptr () -> m ()

Usage: glNamedBufferSubData buffer offset size data

glNamedFramebufferDrawBuffer :: MonadIO m => GLuint -> GLenum -> m ()

Usage: glNamedFramebufferDrawBuffer framebuffer buf

glNamedFramebufferDrawBuffers :: MonadIO m => GLuint -> GLsizei -> Ptr GLenum -> m ()

Usage: glNamedFramebufferDrawBuffers framebuffer n bufs

glNamedFramebufferParameteri :: MonadIO m => GLuint -> GLenum -> GLint -> m ()

Usage: glNamedFramebufferParameteri framebuffer pname param

glNamedFramebufferReadBuffer :: MonadIO m => GLuint -> GLenum -> m ()

Usage: glNamedFramebufferReadBuffer framebuffer src

glNamedFramebufferRenderbuffer :: MonadIO m => GLuint -> GLenum -> GLenum -> GLuint -> m ()

Usage: glNamedFramebufferRenderbuffer framebuffer attachment renderbuffertarget renderbuffer

glNamedFramebufferTexture :: MonadIO m => GLuint -> GLenum -> GLuint -> GLint -> m ()

Usage: glNamedFramebufferTexture framebuffer attachment texture level

glNamedFramebufferTextureLayer :: MonadIO m => GLuint -> GLenum -> GLuint -> GLint -> GLint -> m ()

Usage: glNamedFramebufferTextureLayer framebuffer attachment texture level layer

glNamedRenderbufferStorage :: MonadIO m => GLuint -> GLenum -> GLsizei -> GLsizei -> m ()

Usage: glNamedRenderbufferStorage renderbuffer internalformat width height

glNamedRenderbufferStorageMultisample :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> GLsizei -> m ()

Usage: glNamedRenderbufferStorageMultisample renderbuffer samples internalformat width height

glReadnPixels :: MonadIO m => GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glReadnPixels x y width height format type bufSize data

glTextureBarrier :: MonadIO m => m ()

Usage: glTextureBarrier

Manual page: https://www.opengl.org/sdk/docs/man/html/glTextureBarrier.xhtml

glTextureBuffer :: MonadIO m => GLuint -> GLenum -> GLuint -> m ()

Usage: glTextureBuffer texture internalformat buffer

glTextureBufferRange :: MonadIO m => GLuint -> GLenum -> GLuint -> GLintptr -> GLsizei -> m ()

Usage: glTextureBufferRange texture internalformat buffer offset size

glTextureParameterIiv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glTextureParameterIiv texture pname params

glTextureParameterIuiv :: MonadIO m => GLuint -> GLenum -> Ptr GLuint -> m ()

Usage: glTextureParameterIuiv texture pname params

glTextureParameterf :: MonadIO m => GLuint -> GLenum -> GLfloat -> m ()

Usage: glTextureParameterf texture pname param

glTextureParameterfv :: MonadIO m => GLuint -> GLenum -> Ptr GLfloat -> m ()

Usage: glTextureParameterfv texture pname param

glTextureParameteri :: MonadIO m => GLuint -> GLenum -> GLint -> m ()

Usage: glTextureParameteri texture pname param

glTextureParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glTextureParameteriv texture pname param

glTextureStorage1D :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> m ()

Usage: glTextureStorage1D texture levels internalformat width

glTextureStorage2D :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> GLsizei -> m ()

Usage: glTextureStorage2D texture levels internalformat width height

glTextureStorage2DMultisample :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLboolean -> m ()

Usage: glTextureStorage2DMultisample texture samples internalformat width height fixedsamplelocations

glTextureStorage3D :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLsizei -> m ()

Usage: glTextureStorage3D texture levels internalformat width height depth

glTextureStorage3DMultisample :: MonadIO m => GLuint -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLboolean -> m ()

Usage: glTextureStorage3DMultisample texture samples internalformat width height depth fixedsamplelocations

glTextureSubImage1D :: MonadIO m => GLuint -> GLint -> GLint -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTextureSubImage1D texture level xoffset width format type pixels

glTextureSubImage2D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTextureSubImage2D texture level xoffset yoffset width height format type pixels

glTextureSubImage3D :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTextureSubImage3D texture level xoffset yoffset zoffset width height depth format type pixels

glTransformFeedbackBufferBase :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glTransformFeedbackBufferBase xfb index buffer

Manual page: https://www.opengl.org/sdk/docs/man/html/glTransformFeedbackBufferBase.xhtml

glTransformFeedbackBufferRange :: MonadIO m => GLuint -> GLuint -> GLuint -> GLintptr -> GLsizei -> m ()

Usage: glTransformFeedbackBufferRange xfb index buffer offset size

Manual page: https://www.opengl.org/sdk/docs/man/html/glTransformFeedbackBufferRange.xhtml

glUnmapNamedBuffer :: MonadIO m => GLuint -> m GLboolean

Usage: glUnmapNamedBuffer buffer

glVertexArrayAttribBinding :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glVertexArrayAttribBinding vaobj attribindex bindingindex

glVertexArrayAttribFormat :: MonadIO m => GLuint -> GLuint -> GLint -> GLenum -> GLboolean -> GLuint -> m ()

Usage: glVertexArrayAttribFormat vaobj attribindex size type normalized relativeoffset

glVertexArrayAttribIFormat :: MonadIO m => GLuint -> GLuint -> GLint -> GLenum -> GLuint -> m ()

Usage: glVertexArrayAttribIFormat vaobj attribindex size type relativeoffset

glVertexArrayAttribLFormat :: MonadIO m => GLuint -> GLuint -> GLint -> GLenum -> GLuint -> m ()

Usage: glVertexArrayAttribLFormat vaobj attribindex size type relativeoffset

glVertexArrayBindingDivisor :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glVertexArrayBindingDivisor vaobj bindingindex divisor

glVertexArrayElementBuffer :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glVertexArrayElementBuffer vaobj buffer

Manual page: https://www.opengl.org/sdk/docs/man/html/glVertexArrayElementBuffer.xhtml

glVertexArrayVertexBuffer :: MonadIO m => GLuint -> GLuint -> GLuint -> GLintptr -> GLsizei -> m ()

Usage: glVertexArrayVertexBuffer vaobj bindingindex buffer offset stride

glVertexArrayVertexBuffers :: MonadIO m => GLuint -> GLuint -> GLsizei -> Ptr GLuint -> Ptr GLintptr -> Ptr GLsizei -> m ()

Usage: glVertexArrayVertexBuffers vaobj first count buffers offsets strides

pattern GL_CLIP_DEPTH_MODE :: (Num a, Eq a) => a

pattern GL_CLIP_ORIGIN :: (Num a, Eq a) => a

pattern GL_CONTEXT_FLAG_ROBUST_ACCESS_BIT :: (Num a, Eq a) => a

pattern GL_CONTEXT_LOST :: (Num a, Eq a) => a

pattern GL_CONTEXT_RELEASE_BEHAVIOR :: (Num a, Eq a) => a

pattern GL_CONTEXT_RELEASE_BEHAVIOR_FLUSH :: (Num a, Eq a) => a

pattern GL_GUILTY_CONTEXT_RESET :: (Num a, Eq a) => a

pattern GL_INNOCENT_CONTEXT_RESET :: (Num a, Eq a) => a

pattern GL_LOSE_CONTEXT_ON_RESET :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_CLIP_AND_CULL_DISTANCES :: (Num a, Eq a) => a

pattern GL_MAX_CULL_DISTANCES :: (Num a, Eq a) => a

pattern GL_NEGATIVE_ONE_TO_ONE :: (Num a, Eq a) => a

pattern GL_NO_RESET_NOTIFICATION :: (Num a, Eq a) => a

pattern GL_QUERY_BY_REGION_NO_WAIT_INVERTED :: (Num a, Eq a) => a

pattern GL_QUERY_BY_REGION_WAIT_INVERTED :: (Num a, Eq a) => a

pattern GL_QUERY_NO_WAIT_INVERTED :: (Num a, Eq a) => a

pattern GL_QUERY_TARGET :: (Num a, Eq a) => a

pattern GL_QUERY_WAIT_INVERTED :: (Num a, Eq a) => a

pattern GL_RESET_NOTIFICATION_STRATEGY :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING :: (Num a, Eq a) => a

pattern GL_TEXTURE_TARGET :: (Num a, Eq a) => a

pattern GL_UNKNOWN_CONTEXT_RESET :: (Num a, Eq a) => a

pattern GL_ZERO_TO_ONE :: (Num a, Eq a) => a