Graphics.GL.Raw.Profile.Embedded30

Graphics.GL.Raw.Profile.Embedded30

module Graphics.GL.Raw.Profile.Embedded20

glBeginQuery :: MonadIO m => GLenum -> GLuint -> m ()

Usage: glBeginQuery target id

Manual page: https://www.opengl.org/sdk/docs/man/html/glBeginQuery.xhtml

glBeginTransformFeedback :: MonadIO m => GLenum -> m ()

Usage: glBeginTransformFeedback primitiveMode

Manual page: https://www.opengl.org/sdk/docs/man/html/glBeginTransformFeedback.xhtml

glBindBufferBase :: MonadIO m => GLenum -> GLuint -> GLuint -> m ()

Usage: glBindBufferBase target index buffer

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindBufferBase.xhtml

glBindBufferRange :: MonadIO m => GLenum -> GLuint -> GLuint -> GLintptr -> GLsizeiptr -> m ()

Usage: glBindBufferRange target index buffer offset size

The parameter offset is a BufferOffset.

The parameter size is a BufferSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindBufferRange.xhtml

glBindSampler :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glBindSampler unit sampler

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindSampler.xhtml

glBindTransformFeedback :: MonadIO m => GLenum -> GLuint -> m ()

Usage: glBindTransformFeedback target id

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindTransformFeedback.xhtml

glBindVertexArray :: MonadIO m => GLuint -> m ()

Usage: glBindVertexArray array

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindVertexArray.xhtml

glBlitFramebuffer :: MonadIO m => GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLbitfield -> GLenum -> m ()

Usage: glBlitFramebuffer srcX0 srcY0 srcX1 srcY1 dstX0 dstY0 dstX1 dstY1 mask filter

The parameter mask is a ClearBufferMask, one of: GL_ACCUM_BUFFER_BIT, GL_COLOR_BUFFER_BIT, Graphics.GL.Raw.Extension.NV.CoverageSample.GL_COVERAGE_BUFFER_BIT_NV', GL_DEPTH_BUFFER_BIT, GL_STENCIL_BUFFER_BIT.

Manual page: https://www.opengl.org/sdk/docs/man/html/glBlitFramebuffer.xhtml

glClearBufferfi :: MonadIO m => GLenum -> GLint -> GLfloat -> GLint -> m ()

Usage: glClearBufferfi buffer drawbuffer depth stencil

The parameter drawbuffer is a DrawBufferName.

glClearBufferfv :: MonadIO m => GLenum -> GLint -> Ptr GLfloat -> m ()

Usage: glClearBufferfv buffer drawbuffer value

The parameter drawbuffer is a DrawBufferName.

The length of value should be COMPSIZE(buffer).

glClearBufferiv :: MonadIO m => GLenum -> GLint -> Ptr GLint -> m ()

Usage: glClearBufferiv buffer drawbuffer value

The parameter drawbuffer is a DrawBufferName.

The length of value should be COMPSIZE(buffer).

glClearBufferuiv :: MonadIO m => GLenum -> GLint -> Ptr GLuint -> m ()

Usage: glClearBufferuiv buffer drawbuffer value

The parameter drawbuffer is a DrawBufferName.

The length of value should be COMPSIZE(buffer).

glClientWaitSync :: MonadIO m => GLsync -> GLbitfield -> GLuint64 -> m GLenum

Usage: glClientWaitSync sync flags timeout

The parameter sync is a sync.

Manual page: https://www.opengl.org/sdk/docs/man/html/glClientWaitSync.xhtml

glCompressedTexImage3D :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage3D target level internalformat width height depth border imageSize data

The parameter target is a TextureTarget, one of: Graphics.GL.Raw.Extension.SGIS.DetailTexture.GL_DETAIL_TEXTURE_2D_SGIS', GL_PROXY_TEXTURE_1D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_1D_EXT', GL_PROXY_TEXTURE_2D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_2D_EXT', GL_PROXY_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_PROXY_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_PROXY_TEXTURE_4D_SGIS', GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.OES.Texture3D.GL_TEXTURE_3D_OES', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_TEXTURE_4D_SGIS', GL_TEXTURE_BASE_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_BASE_LEVEL_SGIS', GL_TEXTURE_MAX_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LEVEL_SGIS', GL_TEXTURE_MAX_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LOD_SGIS', GL_TEXTURE_MIN_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MIN_LOD_SGIS'.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexImage3D.xhtml

glCompressedTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage3D target level xoffset yoffset zoffset width height depth format imageSize data

The parameter target is a TextureTarget, one of: Graphics.GL.Raw.Extension.SGIS.DetailTexture.GL_DETAIL_TEXTURE_2D_SGIS', GL_PROXY_TEXTURE_1D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_1D_EXT', GL_PROXY_TEXTURE_2D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_2D_EXT', GL_PROXY_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_PROXY_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_PROXY_TEXTURE_4D_SGIS', GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.OES.Texture3D.GL_TEXTURE_3D_OES', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_TEXTURE_4D_SGIS', GL_TEXTURE_BASE_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_BASE_LEVEL_SGIS', GL_TEXTURE_MAX_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LEVEL_SGIS', GL_TEXTURE_MAX_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LOD_SGIS', GL_TEXTURE_MIN_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MIN_LOD_SGIS'.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: Graphics.GL.Raw.Extension.EXT.Abgr.GL_ABGR_EXT', GL_ALPHA, GL_BLUE, Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYKA_EXT', Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYK_EXT', GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, Graphics.GL.Raw.Extension.EXT.TextureRg.GL_RED_EXT', GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_422_SGIX', Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_444_SGIX'.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexSubImage3D.xhtml

glCopyBufferSubData :: MonadIO m => GLenum -> GLenum -> GLintptr -> GLintptr -> GLsizeiptr -> m ()

Usage: glCopyBufferSubData readTarget writeTarget readOffset writeOffset size

The parameter readOffset is a BufferOffset.

The parameter writeOffset is a BufferOffset.

The parameter size is a BufferSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyBufferSubData.xhtml

glCopyTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage3D target level xoffset yoffset zoffset x y width height

The parameter target is a TextureTarget, one of: Graphics.GL.Raw.Extension.SGIS.DetailTexture.GL_DETAIL_TEXTURE_2D_SGIS', GL_PROXY_TEXTURE_1D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_1D_EXT', GL_PROXY_TEXTURE_2D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_2D_EXT', GL_PROXY_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_PROXY_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_PROXY_TEXTURE_4D_SGIS', GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.OES.Texture3D.GL_TEXTURE_3D_OES', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_TEXTURE_4D_SGIS', GL_TEXTURE_BASE_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_BASE_LEVEL_SGIS', GL_TEXTURE_MAX_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LEVEL_SGIS', GL_TEXTURE_MAX_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LOD_SGIS', GL_TEXTURE_MIN_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MIN_LOD_SGIS'.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexSubImage3D.xhtml

glDeleteQueries :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glDeleteQueries n ids

The length of ids should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteQueries.xhtml

glDeleteSamplers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glDeleteSamplers count samplers

The length of samplers should be count.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteSamplers.xhtml

glDeleteSync :: MonadIO m => GLsync -> m ()

Usage: glDeleteSync sync

The parameter sync is a sync.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteSync.xhtml

glDeleteTransformFeedbacks :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glDeleteTransformFeedbacks n ids

The length of ids should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteTransformFeedbacks.xhtml

glDeleteVertexArrays :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glDeleteVertexArrays n arrays

The length of arrays should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteVertexArrays.xhtml

glDrawArraysInstanced :: MonadIO m => GLenum -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glDrawArraysInstanced mode first count instancecount

The parameter mode is a PrimitiveType, one of: GL_LINES, Graphics.GL.Raw.Profile.Core32.GL_LINES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINES_ADJACENCY_ARB', GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_LINE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINE_STRIP_ADJACENCY_ARB', GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_PATCHES_EXT', GL_POINTS, GL_POLYGON, GL_QUADS, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_QUADS_EXT', GL_QUAD_STRIP, GL_TRIANGLES, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLES_ADJACENCY_ARB', GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLE_STRIP_ADJACENCY_ARB', GL_TRIANGLE_STRIP_ADJACENCY_EXT.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawArraysInstanced.xhtml

glDrawBuffers :: MonadIO m => GLsizei -> Ptr GLenum -> m ()

Usage: glDrawBuffers n bufs

The parameter bufs is a DrawBufferModeATI.

The length of bufs should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawBuffers.xhtml

glDrawElementsInstanced :: MonadIO m => GLenum -> GLsizei -> GLenum -> Ptr () -> GLsizei -> m ()

Usage: glDrawElementsInstanced mode count type indices instancecount

The parameter mode is a PrimitiveType, one of: GL_LINES, Graphics.GL.Raw.Profile.Core32.GL_LINES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINES_ADJACENCY_ARB', GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_LINE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINE_STRIP_ADJACENCY_ARB', GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_PATCHES_EXT', GL_POINTS, GL_POLYGON, GL_QUADS, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_QUADS_EXT', GL_QUAD_STRIP, GL_TRIANGLES, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLES_ADJACENCY_ARB', GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLE_STRIP_ADJACENCY_ARB', GL_TRIANGLE_STRIP_ADJACENCY_EXT.

The parameter type is a DrawElementsType.

The length of indices should be COMPSIZE(count,type).

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawElementsInstanced.xhtml

glDrawRangeElements :: MonadIO m => GLenum -> GLuint -> GLuint -> GLsizei -> GLenum -> Ptr () -> m ()

Usage: glDrawRangeElements mode start end count type indices

The parameter mode is a PrimitiveType, one of: GL_LINES, Graphics.GL.Raw.Profile.Core32.GL_LINES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINES_ADJACENCY_ARB', GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_LINE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_LINE_STRIP_ADJACENCY_ARB', GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_PATCHES_EXT', GL_POINTS, GL_POLYGON, GL_QUADS, Graphics.GL.Raw.Extension.EXT.TessellationShader.GL_QUADS_EXT', GL_QUAD_STRIP, GL_TRIANGLES, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLES_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLES_ADJACENCY_ARB', GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, Graphics.GL.Raw.Profile.Core32.GL_TRIANGLE_STRIP_ADJACENCY', Graphics.GL.Raw.Extension.ARB.GeometryShader4.GL_TRIANGLE_STRIP_ADJACENCY_ARB', GL_TRIANGLE_STRIP_ADJACENCY_EXT.

The parameter type is a DrawElementsType.

The length of indices should be COMPSIZE(count,type).

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawRangeElements.xhtml

glEndQuery :: MonadIO m => GLenum -> m ()

Usage: glEndQuery target

glEndTransformFeedback :: MonadIO m => m ()

Usage: glEndTransformFeedback

glFenceSync :: MonadIO m => GLenum -> GLbitfield -> m GLsync

Usage: glFenceSync condition flags

Manual page: https://www.opengl.org/sdk/docs/man/html/glFenceSync.xhtml

glFlushMappedBufferRange :: MonadIO m => GLenum -> GLintptr -> GLsizeiptr -> m ()

Usage: glFlushMappedBufferRange target offset length

The parameter target is a BufferTargetARB.

The parameter offset is a BufferOffset.

The parameter length is a BufferSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glFlushMappedBufferRange.xhtml

glFramebufferTextureLayer :: MonadIO m => GLenum -> GLenum -> GLuint -> GLint -> GLint -> m ()

Usage: glFramebufferTextureLayer target attachment texture level layer

The parameter target is a FramebufferTarget.

The parameter attachment is a FramebufferAttachment.

The parameter texture is a Texture.

The parameter level is a CheckedInt32.

The parameter layer is a CheckedInt32.

Manual page: https://www.opengl.org/sdk/docs/man/html/glFramebufferTextureLayer.xhtml

glGenQueries :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glGenQueries n ids

The length of ids should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGenQueries.xhtml

glGenSamplers :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glGenSamplers count samplers

The length of samplers should be count.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGenSamplers.xhtml

glGenTransformFeedbacks :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glGenTransformFeedbacks n ids

The length of ids should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGenTransformFeedbacks.xhtml

glGenVertexArrays :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glGenVertexArrays n arrays

The length of arrays should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGenVertexArrays.xhtml

glGetActiveUniformBlockName :: MonadIO m => GLuint -> GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLchar -> m ()

Usage: glGetActiveUniformBlockName program uniformBlockIndex bufSize length uniformBlockName

The length of length should be 1.

The length of uniformBlockName should be bufSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetActiveUniformBlockName.xhtml

glGetActiveUniformBlockiv :: MonadIO m => GLuint -> GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetActiveUniformBlockiv program uniformBlockIndex pname params

The length of params should be COMPSIZE(program,uniformBlockIndex,pname).

glGetActiveUniformsiv :: MonadIO m => GLuint -> GLsizei -> Ptr GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetActiveUniformsiv program uniformCount uniformIndices pname params

The length of uniformIndices should be uniformCount.

The length of params should be COMPSIZE(uniformCount,pname).

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetActiveUniformsiv.xhtml

glGetBufferParameteri64v :: MonadIO m => GLenum -> GLenum -> Ptr GLint64 -> m ()

Usage: glGetBufferParameteri64v target pname params

The parameter target is a BufferTargetARB.

The parameter pname is a BufferPNameARB.

The length of params should be COMPSIZE(pname).

glGetBufferPointerv :: MonadIO m => GLenum -> GLenum -> Ptr (Ptr ()) -> m ()

Usage: glGetBufferPointerv target pname params

The parameter target is a BufferTargetARB.

The parameter pname is a BufferPointerNameARB.

The length of params should be 1.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetBufferPointerv.xhtml

glGetFragDataLocation :: MonadIO m => GLuint -> Ptr GLchar -> m GLint

Usage: glGetFragDataLocation program name

The length of name should be COMPSIZE(name).

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetFragDataLocation.xhtml

glGetInteger64i_v :: MonadIO m => GLenum -> GLuint -> Ptr GLint64 -> m ()

Usage: glGetInteger64i_v target index data

The length of data should be COMPSIZE(target).

glGetInteger64v :: MonadIO m => GLenum -> Ptr GLint64 -> m ()

Usage: glGetInteger64v pname data

The length of data should be COMPSIZE(pname).

glGetIntegeri_v :: MonadIO m => GLenum -> GLuint -> Ptr GLint -> m ()

Usage: glGetIntegeri_v target index data

The length of data should be COMPSIZE(target).

glGetInternalformativ :: MonadIO m => GLenum -> GLenum -> GLenum -> GLsizei -> Ptr GLint -> m ()

Usage: glGetInternalformativ target internalformat pname bufSize params

The length of params should be bufSize.

glGetProgramBinary :: MonadIO m => GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLenum -> Ptr () -> m ()

Usage: glGetProgramBinary program bufSize length binaryFormat binary

The length of length should be 1.

The length of binaryFormat should be 1.

The length of binary should be bufSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetProgramBinary.xhtml

glGetQueryObjectuiv :: MonadIO m => GLuint -> GLenum -> Ptr GLuint -> m ()

Usage: glGetQueryObjectuiv id pname params

The length of params should be COMPSIZE(pname).

glGetQueryiv :: MonadIO m => GLenum -> GLenum -> Ptr GLint -> m ()

Usage: glGetQueryiv target pname params

The length of params should be COMPSIZE(pname).

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetQueryiv.xhtml

glGetSamplerParameterfv :: MonadIO m => GLuint -> GLenum -> Ptr GLfloat -> m ()

Usage: glGetSamplerParameterfv sampler pname params

The length of params should be COMPSIZE(pname).

glGetSamplerParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetSamplerParameteriv sampler pname params

The length of params should be COMPSIZE(pname).

glGetStringi :: MonadIO m => GLenum -> GLuint -> m (Ptr GLubyte)

Usage: glGetStringi name index

glGetSynciv :: MonadIO m => GLsync -> GLenum -> GLsizei -> Ptr GLsizei -> Ptr GLint -> m ()

Usage: glGetSynciv sync pname bufSize length values

The parameter sync is a sync.

The length of length should be 1.

The length of values should be bufSize.

glGetTransformFeedbackVarying :: MonadIO m => GLuint -> GLuint -> GLsizei -> Ptr GLsizei -> Ptr GLsizei -> Ptr GLenum -> Ptr GLchar -> m ()

Usage: glGetTransformFeedbackVarying program index bufSize length size type name

The length of length should be 1.

The length of size should be 1.

The length of type should be 1.

The length of name should be bufSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetTransformFeedbackVarying.xhtml

glGetUniformBlockIndex :: MonadIO m => GLuint -> Ptr GLchar -> m GLuint

Usage: glGetUniformBlockIndex program uniformBlockName

The length of uniformBlockName should be COMPSIZE().

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetUniformBlockIndex.xhtml

glGetUniformIndices :: MonadIO m => GLuint -> GLsizei -> Ptr (Ptr GLchar) -> Ptr GLuint -> m ()

Usage: glGetUniformIndices program uniformCount uniformNames uniformIndices

The length of uniformNames should be COMPSIZE(uniformCount).

The length of uniformIndices should be COMPSIZE(uniformCount).

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetUniformIndices.xhtml

glGetUniformuiv :: MonadIO m => GLuint -> GLint -> Ptr GLuint -> m ()

Usage: glGetUniformuiv program location params

The length of params should be COMPSIZE(program,location).

glGetVertexAttribIiv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glGetVertexAttribIiv index pname params

The parameter pname is a VertexAttribEnum.

The length of params should be 1.

glGetVertexAttribIuiv :: MonadIO m => GLuint -> GLenum -> Ptr GLuint -> m ()

Usage: glGetVertexAttribIuiv index pname params

The parameter pname is a VertexAttribEnum.

The length of params should be 1.

glInvalidateFramebuffer :: MonadIO m => GLenum -> GLsizei -> Ptr GLenum -> m ()

Usage: glInvalidateFramebuffer target numAttachments attachments

The length of attachments should be numAttachments.

Manual page: https://www.opengl.org/sdk/docs/man/html/glInvalidateFramebuffer.xhtml

glInvalidateSubFramebuffer :: MonadIO m => GLenum -> GLsizei -> Ptr GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glInvalidateSubFramebuffer target numAttachments attachments x y width height

The length of attachments should be numAttachments.

Manual page: https://www.opengl.org/sdk/docs/man/html/glInvalidateSubFramebuffer.xhtml

glIsQuery :: MonadIO m => GLuint -> m GLboolean

Usage: glIsQuery id

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsQuery.xhtml

glIsSampler :: MonadIO m => GLuint -> m GLboolean

Usage: glIsSampler sampler

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsSampler.xhtml

glIsSync :: MonadIO m => GLsync -> m GLboolean

Usage: glIsSync sync

The parameter sync is a sync.

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsSync.xhtml

glIsTransformFeedback :: MonadIO m => GLuint -> m GLboolean

Usage: glIsTransformFeedback id

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsTransformFeedback.xhtml

glIsVertexArray :: MonadIO m => GLuint -> m GLboolean

Usage: glIsVertexArray array

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsVertexArray.xhtml

glMapBufferRange :: MonadIO m => GLenum -> GLintptr -> GLsizeiptr -> GLbitfield -> m (Ptr ())

Usage: glMapBufferRange target offset length access

The parameter target is a BufferTargetARB.

The parameter offset is a BufferOffset.

The parameter length is a BufferSize.

The parameter access is a BufferAccessMask.

Manual page: https://www.opengl.org/sdk/docs/man/html/glMapBufferRange.xhtml

glPauseTransformFeedback :: MonadIO m => m ()

Usage: glPauseTransformFeedback

Manual page: https://www.opengl.org/sdk/docs/man/html/glPauseTransformFeedback.xhtml

glProgramBinary :: MonadIO m => GLuint -> GLenum -> Ptr () -> GLsizei -> m ()

Usage: glProgramBinary program binaryFormat binary length

The length of binary should be length.

Manual page: https://www.opengl.org/sdk/docs/man/html/glProgramBinary.xhtml

glProgramParameteri :: MonadIO m => GLuint -> GLenum -> GLint -> m ()

Usage: glProgramParameteri program pname value

The parameter pname is a ProgramParameterPName.

glReadBuffer :: MonadIO m => GLenum -> m ()

Usage: glReadBuffer src

The parameter src is a ReadBufferMode, one of: GL_AUX0, GL_AUX1, GL_AUX2, GL_AUX3, GL_BACK, GL_BACK_LEFT, GL_BACK_RIGHT, GL_FRONT, GL_FRONT_LEFT, GL_FRONT_RIGHT, GL_LEFT, GL_RIGHT.

Manual page: https://www.opengl.org/sdk/docs/man/html/glReadBuffer.xhtml

glRenderbufferStorageMultisample :: MonadIO m => GLenum -> GLsizei -> GLenum -> GLsizei -> GLsizei -> m ()

Usage: glRenderbufferStorageMultisample target samples internalformat width height

Manual page: https://www.opengl.org/sdk/docs/man/html/glRenderbufferStorageMultisample.xhtml

glResumeTransformFeedback :: MonadIO m => m ()

Usage: glResumeTransformFeedback

Manual page: https://www.opengl.org/sdk/docs/man/html/glResumeTransformFeedback.xhtml

glSamplerParameterf :: MonadIO m => GLuint -> GLenum -> GLfloat -> m ()

Usage: glSamplerParameterf sampler pname param

glSamplerParameterfv :: MonadIO m => GLuint -> GLenum -> Ptr GLfloat -> m ()

Usage: glSamplerParameterfv sampler pname param

The length of param should be COMPSIZE(pname).

glSamplerParameteri :: MonadIO m => GLuint -> GLenum -> GLint -> m ()

Usage: glSamplerParameteri sampler pname param

glSamplerParameteriv :: MonadIO m => GLuint -> GLenum -> Ptr GLint -> m ()

Usage: glSamplerParameteriv sampler pname param

The length of param should be COMPSIZE(pname).

glTexImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexImage3D target level internalformat width height depth border format type pixels

The parameter target is a TextureTarget, one of: Graphics.GL.Raw.Extension.SGIS.DetailTexture.GL_DETAIL_TEXTURE_2D_SGIS', GL_PROXY_TEXTURE_1D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_1D_EXT', GL_PROXY_TEXTURE_2D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_2D_EXT', GL_PROXY_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_PROXY_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_PROXY_TEXTURE_4D_SGIS', GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.OES.Texture3D.GL_TEXTURE_3D_OES', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_TEXTURE_4D_SGIS', GL_TEXTURE_BASE_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_BASE_LEVEL_SGIS', GL_TEXTURE_MAX_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LEVEL_SGIS', GL_TEXTURE_MAX_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LOD_SGIS', GL_TEXTURE_MIN_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MIN_LOD_SGIS'.

The parameter level is a CheckedInt32.

The parameter internalformat is a TextureComponentCount.

The parameter border is a CheckedInt32.

The parameter format is a PixelFormat, one of: Graphics.GL.Raw.Extension.EXT.Abgr.GL_ABGR_EXT', GL_ALPHA, GL_BLUE, Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYKA_EXT', Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYK_EXT', GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, Graphics.GL.Raw.Extension.EXT.TextureRg.GL_RED_EXT', GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_422_SGIX', Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_444_SGIX'.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_BYTE_3_3_2_EXT', GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_INT_10_10_10_2_EXT', GL_UNSIGNED_INT_8_8_8_8, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_INT_8_8_8_8_EXT', GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_SHORT_4_4_4_4_EXT', GL_UNSIGNED_SHORT_5_5_5_1, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_SHORT_5_5_5_1_EXT'.

The length of pixels should be COMPSIZE(format,type,width,height,depth).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexImage3D.xhtml

glTexStorage2D :: MonadIO m => GLenum -> GLsizei -> GLenum -> GLsizei -> GLsizei -> m ()

Usage: glTexStorage2D target levels internalformat width height

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexStorage2D.xhtml

glTexStorage3D :: MonadIO m => GLenum -> GLsizei -> GLenum -> GLsizei -> GLsizei -> GLsizei -> m ()

Usage: glTexStorage3D target levels internalformat width height depth

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexStorage3D.xhtml

glTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexSubImage3D target level xoffset yoffset zoffset width height depth format type pixels

The parameter target is a TextureTarget, one of: Graphics.GL.Raw.Extension.SGIS.DetailTexture.GL_DETAIL_TEXTURE_2D_SGIS', GL_PROXY_TEXTURE_1D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_1D_EXT', GL_PROXY_TEXTURE_2D, Graphics.GL.Raw.Extension.EXT.Texture.GL_PROXY_TEXTURE_2D_EXT', GL_PROXY_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_PROXY_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_PROXY_TEXTURE_4D_SGIS', GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, Graphics.GL.Raw.Extension.EXT.Texture3D.GL_TEXTURE_3D_EXT', Graphics.GL.Raw.Extension.OES.Texture3D.GL_TEXTURE_3D_OES', Graphics.GL.Raw.Extension.SGIS.Texture4D.GL_TEXTURE_4D_SGIS', GL_TEXTURE_BASE_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_BASE_LEVEL_SGIS', GL_TEXTURE_MAX_LEVEL, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LEVEL_SGIS', GL_TEXTURE_MAX_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MAX_LOD_SGIS', GL_TEXTURE_MIN_LOD, Graphics.GL.Raw.Extension.SGIS.TextureLod.GL_TEXTURE_MIN_LOD_SGIS'.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: Graphics.GL.Raw.Extension.EXT.Abgr.GL_ABGR_EXT', GL_ALPHA, GL_BLUE, Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYKA_EXT', Graphics.GL.Raw.Extension.EXT.Cmyka.GL_CMYK_EXT', GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, Graphics.GL.Raw.Extension.EXT.TextureRg.GL_RED_EXT', GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_422_SGIX', Graphics.GL.Raw.Extension.SGIX.Ycrcb.GL_YCRCB_444_SGIX'.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_BYTE_3_3_2_EXT', GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_INT_10_10_10_2_EXT', GL_UNSIGNED_INT_8_8_8_8, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_INT_8_8_8_8_EXT', GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_SHORT_4_4_4_4_EXT', GL_UNSIGNED_SHORT_5_5_5_1, Graphics.GL.Raw.Extension.EXT.PackedPixels.GL_UNSIGNED_SHORT_5_5_5_1_EXT'.

The length of pixels should be COMPSIZE(format,type,width,height,depth).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexSubImage3D.xhtml

glTransformFeedbackVaryings :: MonadIO m => GLuint -> GLsizei -> Ptr (Ptr GLchar) -> GLenum -> m ()

Usage: glTransformFeedbackVaryings program count varyings bufferMode

The length of varyings should be count.

Manual page: https://www.opengl.org/sdk/docs/man/html/glTransformFeedbackVaryings.xhtml

glUniform1ui :: MonadIO m => GLint -> GLuint -> m ()

Usage: glUniform1ui location v0

glUniform1uiv :: MonadIO m => GLint -> GLsizei -> Ptr GLuint -> m ()

Usage: glUniform1uiv location count value

The length of value should be count*1.

glUniform2ui :: MonadIO m => GLint -> GLuint -> GLuint -> m ()

Usage: glUniform2ui location v0 v1

glUniform2uiv :: MonadIO m => GLint -> GLsizei -> Ptr GLuint -> m ()

Usage: glUniform2uiv location count value

The length of value should be count*2.

glUniform3ui :: MonadIO m => GLint -> GLuint -> GLuint -> GLuint -> m ()

Usage: glUniform3ui location v0 v1 v2

glUniform3uiv :: MonadIO m => GLint -> GLsizei -> Ptr GLuint -> m ()

Usage: glUniform3uiv location count value

The length of value should be count*3.

glUniform4ui :: MonadIO m => GLint -> GLuint -> GLuint -> GLuint -> GLuint -> m ()

Usage: glUniform4ui location v0 v1 v2 v3

glUniform4uiv :: MonadIO m => GLint -> GLsizei -> Ptr GLuint -> m ()

Usage: glUniform4uiv location count value

The length of value should be count*4.

glUniformBlockBinding :: MonadIO m => GLuint -> GLuint -> GLuint -> m ()

Usage: glUniformBlockBinding program uniformBlockIndex uniformBlockBinding

Manual page: https://www.opengl.org/sdk/docs/man/html/glUniformBlockBinding.xhtml

glUniformMatrix2x3fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix2x3fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*6.

glUniformMatrix2x4fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix2x4fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*8.

glUniformMatrix3x2fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix3x2fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*6.

glUniformMatrix3x4fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix3x4fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*12.

glUniformMatrix4x2fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix4x2fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*8.

glUniformMatrix4x3fv :: MonadIO m => GLint -> GLsizei -> GLboolean -> Ptr GLfloat -> m ()

Usage: glUniformMatrix4x3fv location count transpose value

The parameter transpose is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of value should be count*12.

glUnmapBuffer :: MonadIO m => GLenum -> m GLboolean

Usage: glUnmapBuffer target

The parameter target is a BufferTargetARB.

Manual page: https://www.opengl.org/sdk/docs/man/html/glUnmapBuffer.xhtml

glVertexAttribDivisor :: MonadIO m => GLuint -> GLuint -> m ()

Usage: glVertexAttribDivisor index divisor

Manual page: https://www.opengl.org/sdk/docs/man/html/glVertexAttribDivisor.xhtml

glVertexAttribI4i :: MonadIO m => GLuint -> GLint -> GLint -> GLint -> GLint -> m ()

Usage: glVertexAttribI4i index x y z w

The vector equivalent of this command is glVertexAttribI4iv.

glVertexAttribI4iv :: MonadIO m => GLuint -> Ptr GLint -> m ()

Usage: glVertexAttribI4iv index v

The length of v should be 4.

glVertexAttribI4ui :: MonadIO m => GLuint -> GLuint -> GLuint -> GLuint -> GLuint -> m ()

Usage: glVertexAttribI4ui index x y z w

The vector equivalent of this command is glVertexAttribI4uiv.

glVertexAttribI4uiv :: MonadIO m => GLuint -> Ptr GLuint -> m ()

Usage: glVertexAttribI4uiv index v

The length of v should be 4.

glVertexAttribIPointer :: MonadIO m => GLuint -> GLint -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glVertexAttribIPointer index size type stride pointer

The parameter type is a VertexAttribEnum.

The length of pointer should be COMPSIZE(size,type,stride).

glWaitSync :: MonadIO m => GLsync -> GLbitfield -> GLuint64 -> m ()

Usage: glWaitSync sync flags timeout

The parameter sync is a sync.

Manual page: https://www.opengl.org/sdk/docs/man/html/glWaitSync.xhtml

pattern GL_ACTIVE_UNIFORM_BLOCKS :: (Num a, Eq a) => a

pattern GL_ACTIVE_UNIFORM_BLOCK_MAX_NAME_LENGTH :: (Num a, Eq a) => a

pattern GL_ALREADY_SIGNALED :: (Num a, Eq a) => a

pattern GL_ANY_SAMPLES_PASSED :: (Num a, Eq a) => a

pattern GL_ANY_SAMPLES_PASSED_CONSERVATIVE :: (Num a, Eq a) => a

pattern GL_BLUE :: (Num a, Eq a) => a

pattern GL_BUFFER_ACCESS_FLAGS :: (Num a, Eq a) => a

pattern GL_BUFFER_MAPPED :: (Num a, Eq a) => a

pattern GL_BUFFER_MAP_LENGTH :: (Num a, Eq a) => a

pattern GL_BUFFER_MAP_OFFSET :: (Num a, Eq a) => a

pattern GL_BUFFER_MAP_POINTER :: (Num a, Eq a) => a

pattern GL_COLOR :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT1 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT10 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT11 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT12 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT13 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT14 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT15 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT2 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT3 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT4 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT5 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT6 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT7 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT8 :: (Num a, Eq a) => a

pattern GL_COLOR_ATTACHMENT9 :: (Num a, Eq a) => a

pattern GL_COMPARE_REF_TO_TEXTURE :: (Num a, Eq a) => a

pattern GL_COMPRESSED_R11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RG11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB8_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB8_PUNCHTHROUGH_ALPHA1_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGBA8_ETC2_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SIGNED_R11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SIGNED_RG11_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_ALPHA8_ETC2_EAC :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_ETC2 :: (Num a, Eq a) => a

pattern GL_COMPRESSED_SRGB8_PUNCHTHROUGH_ALPHA1_ETC2 :: (Num a, Eq a) => a

pattern GL_CONDITION_SATISFIED :: (Num a, Eq a) => a

pattern GL_COPY_READ_BUFFER :: (Num a, Eq a) => a

pattern GL_COPY_READ_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_COPY_WRITE_BUFFER :: (Num a, Eq a) => a

pattern GL_COPY_WRITE_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_CURRENT_QUERY :: (Num a, Eq a) => a

pattern GL_DEPTH :: (Num a, Eq a) => a

pattern GL_DEPTH24_STENCIL8 :: (Num a, Eq a) => a

pattern GL_DEPTH32F_STENCIL8 :: (Num a, Eq a) => a

pattern GL_DEPTH_COMPONENT24 :: (Num a, Eq a) => a

pattern GL_DEPTH_COMPONENT32F :: (Num a, Eq a) => a

pattern GL_DEPTH_STENCIL :: (Num a, Eq a) => a

pattern GL_DEPTH_STENCIL_ATTACHMENT :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER0 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER1 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER10 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER11 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER12 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER13 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER14 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER15 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER2 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER3 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER4 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER5 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER6 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER7 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER8 :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER9 :: (Num a, Eq a) => a

pattern GL_DRAW_FRAMEBUFFER :: (Num a, Eq a) => a

pattern GL_DRAW_FRAMEBUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_DYNAMIC_COPY :: (Num a, Eq a) => a

pattern GL_DYNAMIC_READ :: (Num a, Eq a) => a

pattern GL_FLOAT_32_UNSIGNED_INT_24_8_REV :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT2x3 :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT2x4 :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT3x2 :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT3x4 :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT4x2 :: (Num a, Eq a) => a

pattern GL_FLOAT_MAT4x3 :: (Num a, Eq a) => a

pattern GL_FRAGMENT_SHADER_DERIVATIVE_HINT :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_ALPHA_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_BLUE_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_DEPTH_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_GREEN_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_RED_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_STENCIL_SIZE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_ATTACHMENT_TEXTURE_LAYER :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_DEFAULT :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_INCOMPLETE_MULTISAMPLE :: (Num a, Eq a) => a

pattern GL_FRAMEBUFFER_UNDEFINED :: (Num a, Eq a) => a

pattern GL_GREEN :: (Num a, Eq a) => a

pattern GL_HALF_FLOAT :: (Num a, Eq a) => a

pattern GL_INTERLEAVED_ATTRIBS :: (Num a, Eq a) => a

pattern GL_INT_2_10_10_10_REV :: (Num a, Eq a) => a

pattern GL_INT_SAMPLER_2D :: (Num a, Eq a) => a

pattern GL_INT_SAMPLER_2D_ARRAY :: (Num a, Eq a) => a

pattern GL_INT_SAMPLER_3D :: (Num a, Eq a) => a

pattern GL_INT_SAMPLER_CUBE :: (Num a, Eq a) => a

pattern GL_INVALID_INDEX :: (Num a, Eq a) => a

pattern GL_MAJOR_VERSION :: (Num a, Eq a) => a

pattern GL_MAP_FLUSH_EXPLICIT_BIT :: (Num a, Eq a) => a

pattern GL_MAP_INVALIDATE_BUFFER_BIT :: (Num a, Eq a) => a

pattern GL_MAP_INVALIDATE_RANGE_BIT :: (Num a, Eq a) => a

pattern GL_MAP_READ_BIT :: (Num a, Eq a) => a

pattern GL_MAP_UNSYNCHRONIZED_BIT :: (Num a, Eq a) => a

pattern GL_MAP_WRITE_BIT :: (Num a, Eq a) => a

pattern GL_MAX :: (Num a, Eq a) => a

pattern GL_MAX_3D_TEXTURE_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_ARRAY_TEXTURE_LAYERS :: (Num a, Eq a) => a

pattern GL_MAX_COLOR_ATTACHMENTS :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_FRAGMENT_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_UNIFORM_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_COMBINED_VERTEX_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_DRAW_BUFFERS :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENTS_INDICES :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENTS_VERTICES :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENT_INDEX :: (Num a, Eq a) => a

pattern GL_MAX_FRAGMENT_INPUT_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_FRAGMENT_UNIFORM_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_FRAGMENT_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_PROGRAM_TEXEL_OFFSET :: (Num a, Eq a) => a

pattern GL_MAX_SAMPLES :: (Num a, Eq a) => a

pattern GL_MAX_SERVER_WAIT_TIMEOUT :: (Num a, Eq a) => a

pattern GL_MAX_TEXTURE_LOD_BIAS :: (Num a, Eq a) => a

pattern GL_MAX_TRANSFORM_FEEDBACK_INTERLEAVED_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_TRANSFORM_FEEDBACK_SEPARATE_ATTRIBS :: (Num a, Eq a) => a

pattern GL_MAX_TRANSFORM_FEEDBACK_SEPARATE_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_UNIFORM_BLOCK_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_UNIFORM_BUFFER_BINDINGS :: (Num a, Eq a) => a

pattern GL_MAX_VARYING_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_OUTPUT_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_UNIFORM_BLOCKS :: (Num a, Eq a) => a

pattern GL_MAX_VERTEX_UNIFORM_COMPONENTS :: (Num a, Eq a) => a

pattern GL_MIN :: (Num a, Eq a) => a

pattern GL_MINOR_VERSION :: (Num a, Eq a) => a

pattern GL_MIN_PROGRAM_TEXEL_OFFSET :: (Num a, Eq a) => a

pattern GL_NUM_EXTENSIONS :: (Num a, Eq a) => a

pattern GL_NUM_PROGRAM_BINARY_FORMATS :: (Num a, Eq a) => a

pattern GL_NUM_SAMPLE_COUNTS :: (Num a, Eq a) => a

pattern GL_OBJECT_TYPE :: (Num a, Eq a) => a

pattern GL_PACK_ROW_LENGTH :: (Num a, Eq a) => a

pattern GL_PACK_SKIP_PIXELS :: (Num a, Eq a) => a

pattern GL_PACK_SKIP_ROWS :: (Num a, Eq a) => a

pattern GL_PIXEL_PACK_BUFFER :: (Num a, Eq a) => a

pattern GL_PIXEL_PACK_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_PIXEL_UNPACK_BUFFER :: (Num a, Eq a) => a

pattern GL_PIXEL_UNPACK_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_PRIMITIVE_RESTART_FIXED_INDEX :: (Num a, Eq a) => a

pattern GL_PROGRAM_BINARY_FORMATS :: (Num a, Eq a) => a

pattern GL_PROGRAM_BINARY_LENGTH :: (Num a, Eq a) => a

pattern GL_PROGRAM_BINARY_RETRIEVABLE_HINT :: (Num a, Eq a) => a

pattern GL_QUERY_RESULT :: (Num a, Eq a) => a

pattern GL_QUERY_RESULT_AVAILABLE :: (Num a, Eq a) => a

pattern GL_R11F_G11F_B10F :: (Num a, Eq a) => a

pattern GL_R16F :: (Num a, Eq a) => a

pattern GL_R16I :: (Num a, Eq a) => a

pattern GL_R16UI :: (Num a, Eq a) => a

pattern GL_R32F :: (Num a, Eq a) => a

pattern GL_R32I :: (Num a, Eq a) => a

pattern GL_R32UI :: (Num a, Eq a) => a

pattern GL_R8 :: (Num a, Eq a) => a

pattern GL_R8I :: (Num a, Eq a) => a

pattern GL_R8UI :: (Num a, Eq a) => a

pattern GL_R8_SNORM :: (Num a, Eq a) => a

pattern GL_RASTERIZER_DISCARD :: (Num a, Eq a) => a

pattern GL_READ_BUFFER :: (Num a, Eq a) => a

pattern GL_READ_FRAMEBUFFER :: (Num a, Eq a) => a

pattern GL_READ_FRAMEBUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_RED :: (Num a, Eq a) => a

pattern GL_RED_INTEGER :: (Num a, Eq a) => a

pattern GL_RENDERBUFFER_SAMPLES :: (Num a, Eq a) => a

pattern GL_RG :: (Num a, Eq a) => a

pattern GL_RG16F :: (Num a, Eq a) => a

pattern GL_RG16I :: (Num a, Eq a) => a

pattern GL_RG16UI :: (Num a, Eq a) => a

pattern GL_RG32F :: (Num a, Eq a) => a

pattern GL_RG32I :: (Num a, Eq a) => a

pattern GL_RG32UI :: (Num a, Eq a) => a

pattern GL_RG8 :: (Num a, Eq a) => a

pattern GL_RG8I :: (Num a, Eq a) => a

pattern GL_RG8UI :: (Num a, Eq a) => a

pattern GL_RG8_SNORM :: (Num a, Eq a) => a

pattern GL_RGB10_A2 :: (Num a, Eq a) => a

pattern GL_RGB10_A2UI :: (Num a, Eq a) => a

pattern GL_RGB16F :: (Num a, Eq a) => a

pattern GL_RGB16I :: (Num a, Eq a) => a

pattern GL_RGB16UI :: (Num a, Eq a) => a

pattern GL_RGB32F :: (Num a, Eq a) => a

pattern GL_RGB32I :: (Num a, Eq a) => a

pattern GL_RGB32UI :: (Num a, Eq a) => a

pattern GL_RGB8 :: (Num a, Eq a) => a

pattern GL_RGB8I :: (Num a, Eq a) => a

pattern GL_RGB8UI :: (Num a, Eq a) => a

pattern GL_RGB8_SNORM :: (Num a, Eq a) => a

pattern GL_RGB9_E5 :: (Num a, Eq a) => a

pattern GL_RGBA16F :: (Num a, Eq a) => a

pattern GL_RGBA16I :: (Num a, Eq a) => a

pattern GL_RGBA16UI :: (Num a, Eq a) => a

pattern GL_RGBA32F :: (Num a, Eq a) => a

pattern GL_RGBA32I :: (Num a, Eq a) => a

pattern GL_RGBA32UI :: (Num a, Eq a) => a

pattern GL_RGBA8 :: (Num a, Eq a) => a

pattern GL_RGBA8I :: (Num a, Eq a) => a

pattern GL_RGBA8UI :: (Num a, Eq a) => a

pattern GL_RGBA8_SNORM :: (Num a, Eq a) => a

pattern GL_RGBA_INTEGER :: (Num a, Eq a) => a

pattern GL_RGB_INTEGER :: (Num a, Eq a) => a

pattern GL_RG_INTEGER :: (Num a, Eq a) => a

pattern GL_SAMPLER_2D_ARRAY :: (Num a, Eq a) => a

pattern GL_SAMPLER_2D_ARRAY_SHADOW :: (Num a, Eq a) => a

pattern GL_SAMPLER_2D_SHADOW :: (Num a, Eq a) => a

pattern GL_SAMPLER_3D :: (Num a, Eq a) => a

pattern GL_SAMPLER_BINDING :: (Num a, Eq a) => a

pattern GL_SAMPLER_CUBE_SHADOW :: (Num a, Eq a) => a

pattern GL_SEPARATE_ATTRIBS :: (Num a, Eq a) => a

pattern GL_SIGNALED :: (Num a, Eq a) => a

pattern GL_SIGNED_NORMALIZED :: (Num a, Eq a) => a

pattern GL_SRGB :: (Num a, Eq a) => a

pattern GL_SRGB8 :: (Num a, Eq a) => a

pattern GL_SRGB8_ALPHA8 :: (Num a, Eq a) => a

pattern GL_STATIC_COPY :: (Num a, Eq a) => a

pattern GL_STATIC_READ :: (Num a, Eq a) => a

pattern GL_STENCIL :: (Num a, Eq a) => a

pattern GL_STREAM_COPY :: (Num a, Eq a) => a

pattern GL_STREAM_READ :: (Num a, Eq a) => a

pattern GL_SYNC_CONDITION :: (Num a, Eq a) => a

pattern GL_SYNC_FENCE :: (Num a, Eq a) => a

pattern GL_SYNC_FLAGS :: (Num a, Eq a) => a

pattern GL_SYNC_FLUSH_COMMANDS_BIT :: (Num a, Eq a) => a

pattern GL_SYNC_GPU_COMMANDS_COMPLETE :: (Num a, Eq a) => a

pattern GL_SYNC_STATUS :: (Num a, Eq a) => a

pattern GL_TEXTURE_2D_ARRAY :: (Num a, Eq a) => a

pattern GL_TEXTURE_3D :: (Num a, Eq a) => a

pattern GL_TEXTURE_BASE_LEVEL :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_2D_ARRAY :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_3D :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPARE_FUNC :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPARE_MODE :: (Num a, Eq a) => a

pattern GL_TEXTURE_IMMUTABLE_FORMAT :: (Num a, Eq a) => a

pattern GL_TEXTURE_IMMUTABLE_LEVELS :: (Num a, Eq a) => a

pattern GL_TEXTURE_MAX_LEVEL :: (Num a, Eq a) => a

pattern GL_TEXTURE_MAX_LOD :: (Num a, Eq a) => a

pattern GL_TEXTURE_MIN_LOD :: (Num a, Eq a) => a

pattern GL_TEXTURE_SWIZZLE_A :: (Num a, Eq a) => a

pattern GL_TEXTURE_SWIZZLE_B :: (Num a, Eq a) => a

pattern GL_TEXTURE_SWIZZLE_G :: (Num a, Eq a) => a

pattern GL_TEXTURE_SWIZZLE_R :: (Num a, Eq a) => a

pattern GL_TEXTURE_WRAP_R :: (Num a, Eq a) => a

pattern GL_TIMEOUT_EXPIRED :: (Num a, Eq a) => a

pattern GL_TIMEOUT_IGNORED :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_ACTIVE :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BINDING :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BUFFER :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BUFFER_MODE :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_BUFFER_START :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_PAUSED :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_PRIMITIVES_WRITTEN :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_VARYINGS :: (Num a, Eq a) => a

pattern GL_TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH :: (Num a, Eq a) => a

pattern GL_UNIFORM_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_ACTIVE_UNIFORMS :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_BINDING :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_DATA_SIZE :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_INDEX :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_NAME_LENGTH :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_REFERENCED_BY_FRAGMENT_SHADER :: (Num a, Eq a) => a

pattern GL_UNIFORM_BLOCK_REFERENCED_BY_VERTEX_SHADER :: (Num a, Eq a) => a

pattern GL_UNIFORM_BUFFER :: (Num a, Eq a) => a

pattern GL_UNIFORM_BUFFER_BINDING :: (Num a, Eq a) => a

pattern GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT :: (Num a, Eq a) => a

pattern GL_UNIFORM_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_UNIFORM_BUFFER_START :: (Num a, Eq a) => a

pattern GL_UNIFORM_IS_ROW_MAJOR :: (Num a, Eq a) => a

pattern GL_UNIFORM_MATRIX_STRIDE :: (Num a, Eq a) => a

pattern GL_UNIFORM_NAME_LENGTH :: (Num a, Eq a) => a

pattern GL_UNIFORM_OFFSET :: (Num a, Eq a) => a

pattern GL_UNIFORM_SIZE :: (Num a, Eq a) => a

pattern GL_UNIFORM_TYPE :: (Num a, Eq a) => a

pattern GL_UNPACK_IMAGE_HEIGHT :: (Num a, Eq a) => a

pattern GL_UNPACK_ROW_LENGTH :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_IMAGES :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_PIXELS :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_ROWS :: (Num a, Eq a) => a

pattern GL_UNSIGNALED :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_10F_11F_11F_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_24_8 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_2_10_10_10_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_5_9_9_9_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_SAMPLER_2D :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_SAMPLER_2D_ARRAY :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_SAMPLER_3D :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_SAMPLER_CUBE :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_VEC2 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_VEC3 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_VEC4 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_NORMALIZED :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY_BINDING :: (Num a, Eq a) => a

pattern GL_VERTEX_ATTRIB_ARRAY_DIVISOR :: (Num a, Eq a) => a

pattern GL_VERTEX_ATTRIB_ARRAY_INTEGER :: (Num a, Eq a) => a

pattern GL_WAIT_FAILED :: (Num a, Eq a) => a