Graphics.GL.Raw.Profile.Standard11

Graphics.GL.Raw.Profile.Standard11

module Graphics.GL.Raw.Profile.Standard10

glAreTexturesResident :: MonadIO m => GLsizei -> Ptr GLuint -> Ptr GLboolean -> m GLboolean

Usage: glAreTexturesResident n textures residences

The parameter textures is a Texture.

The parameter residences is a Boolean, one of: GL_FALSE, GL_TRUE.

The length of textures should be n.

The length of residences should be n.

glArrayElement :: MonadIO m => GLint -> m ()

Usage: glArrayElement i

glBindTexture :: MonadIO m => GLenum -> GLuint -> m ()

Usage: glBindTexture target texture

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter texture is a Texture.

Manual page: https://www.opengl.org/sdk/docs/man/html/glBindTexture.xhtml

glColorPointer :: MonadIO m => GLint -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glColorPointer size type stride pointer

The parameter type is a ColorPointerType, one of: GL_BYTE, GL_DOUBLE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT.

The length of pointer should be COMPSIZE(size,type,stride).

glCopyTexImage1D :: MonadIO m => GLenum -> GLint -> GLenum -> GLint -> GLint -> GLsizei -> GLint -> m ()

Usage: glCopyTexImage1D target level internalformat x y width border

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

The parameter border is a CheckedInt32.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexImage1D.xhtml

glCopyTexImage2D :: MonadIO m => GLenum -> GLint -> GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> GLint -> m ()

Usage: glCopyTexImage2D target level internalformat x y width height border

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

The parameter border is a CheckedInt32.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexImage2D.xhtml

glCopyTexSubImage1D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> m ()

Usage: glCopyTexSubImage1D target level xoffset x y width

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexSubImage1D.xhtml

glCopyTexSubImage2D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage2D target level xoffset yoffset x y width height

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexSubImage2D.xhtml

glDeleteTextures :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glDeleteTextures n textures

The parameter textures is a Texture.

The length of textures should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDeleteTextures.xhtml

glDisableClientState :: MonadIO m => GLenum -> m ()

Usage: glDisableClientState array

The parameter array is a EnableCap, one of: GL_ALPHA_TEST, GL_ASYNC_DRAW_PIXELS_SGIX, GL_ASYNC_HISTOGRAM_SGIX, GL_ASYNC_READ_PIXELS_SGIX, GL_ASYNC_TEX_IMAGE_SGIX, GL_AUTO_NORMAL, GL_BLEND, GL_CALLIGRAPHIC_FRAGMENT_SGIX, GL_CLIP_PLANE0, GL_CLIP_PLANE1, GL_CLIP_PLANE2, GL_CLIP_PLANE3, GL_CLIP_PLANE4, GL_CLIP_PLANE5, GL_COLOR_ARRAY, GL_COLOR_LOGIC_OP, GL_COLOR_MATERIAL, GL_COLOR_TABLE_SGI, GL_CONVOLUTION_1D_EXT, GL_CONVOLUTION_2D_EXT, GL_CULL_FACE, GL_DEPTH_TEST, GL_DITHER, GL_EDGE_FLAG_ARRAY, GL_FOG, GL_FOG_OFFSET_SGIX, GL_FRAGMENT_COLOR_MATERIAL_SGIX, GL_FRAGMENT_LIGHT0_SGIX, GL_FRAGMENT_LIGHT1_SGIX, GL_FRAGMENT_LIGHT2_SGIX, GL_FRAGMENT_LIGHT3_SGIX, GL_FRAGMENT_LIGHT4_SGIX, GL_FRAGMENT_LIGHT5_SGIX, GL_FRAGMENT_LIGHT6_SGIX, GL_FRAGMENT_LIGHT7_SGIX, GL_FRAGMENT_LIGHTING_SGIX, GL_FRAMEZOOM_SGIX, GL_HISTOGRAM_EXT, GL_INDEX_ARRAY, GL_INDEX_LOGIC_OP, GL_INTERLACE_SGIX, GL_IR_INSTRUMENT1_SGIX, GL_LIGHT0, GL_LIGHT1, GL_LIGHT2, GL_LIGHT3, GL_LIGHT4, GL_LIGHT5, GL_LIGHT6, GL_LIGHT7, GL_LIGHTING, GL_LINE_SMOOTH, GL_LINE_STIPPLE, GL_MAP1_COLOR_4, GL_MAP1_INDEX, GL_MAP1_NORMAL, GL_MAP1_TEXTURE_COORD_1, GL_MAP1_TEXTURE_COORD_2, GL_MAP1_TEXTURE_COORD_3, GL_MAP1_TEXTURE_COORD_4, GL_MAP1_VERTEX_3, GL_MAP1_VERTEX_4, GL_MAP2_COLOR_4, GL_MAP2_INDEX, GL_MAP2_NORMAL, GL_MAP2_TEXTURE_COORD_1, GL_MAP2_TEXTURE_COORD_2, GL_MAP2_TEXTURE_COORD_3, GL_MAP2_TEXTURE_COORD_4, GL_MAP2_VERTEX_3, GL_MAP2_VERTEX_4, GL_MINMAX_EXT, GL_MULTISAMPLE_SGIS, GL_NORMALIZE, GL_NORMAL_ARRAY, GL_PIXEL_TEXTURE_SGIS, GL_PIXEL_TEX_GEN_SGIX, GL_POINT_SMOOTH, GL_POLYGON_OFFSET_FILL, GL_POLYGON_OFFSET_LINE, GL_POLYGON_OFFSET_POINT, GL_POLYGON_SMOOTH, GL_POLYGON_STIPPLE, GL_POST_COLOR_MATRIX_COLOR_TABLE_SGI, GL_POST_CONVOLUTION_COLOR_TABLE_SGI, GL_REFERENCE_PLANE_SGIX, GL_RESCALE_NORMAL_EXT, GL_SAMPLE_ALPHA_TO_MASK_SGIS, GL_SAMPLE_ALPHA_TO_ONE_SGIS, GL_SAMPLE_MASK_SGIS, GL_SCISSOR_TEST, GL_SEPARABLE_2D_EXT, GL_SHARED_TEXTURE_PALETTE_EXT, GL_SPRITE_SGIX, GL_STENCIL_TEST, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D_EXT, GL_TEXTURE_4D_SGIS, GL_TEXTURE_COLOR_TABLE_SGI, GL_TEXTURE_COORD_ARRAY, GL_TEXTURE_GEN_Q, GL_TEXTURE_GEN_R, GL_TEXTURE_GEN_S, GL_TEXTURE_GEN_T, GL_VERTEX_ARRAY.

glDrawArrays :: MonadIO m => GLenum -> GLint -> GLsizei -> m ()

Usage: glDrawArrays mode first count

The parameter mode is a PrimitiveType, one of: GL_LINES, GL_LINES_ADJACENCY, GL_LINES_ADJACENCY_ARB, GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, GL_LINE_STRIP_ADJACENCY, GL_LINE_STRIP_ADJACENCY_ARB, GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, GL_PATCHES_EXT, GL_POINTS, GL_POLYGON, GL_QUADS, GL_QUADS_EXT, GL_QUAD_STRIP, GL_TRIANGLES, GL_TRIANGLES_ADJACENCY, GL_TRIANGLES_ADJACENCY_ARB, GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, GL_TRIANGLE_STRIP_ADJACENCY, GL_TRIANGLE_STRIP_ADJACENCY_ARB, GL_TRIANGLE_STRIP_ADJACENCY_EXT.

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawArrays.xhtml

glDrawElements :: MonadIO m => GLenum -> GLsizei -> GLenum -> Ptr () -> m ()

Usage: glDrawElements mode count type indices

The parameter mode is a PrimitiveType, one of: GL_LINES, GL_LINES_ADJACENCY, GL_LINES_ADJACENCY_ARB, GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, GL_LINE_STRIP_ADJACENCY, GL_LINE_STRIP_ADJACENCY_ARB, GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, GL_PATCHES_EXT, GL_POINTS, GL_POLYGON, GL_QUADS, GL_QUADS_EXT, GL_QUAD_STRIP, GL_TRIANGLES, GL_TRIANGLES_ADJACENCY, GL_TRIANGLES_ADJACENCY_ARB, GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, GL_TRIANGLE_STRIP_ADJACENCY, GL_TRIANGLE_STRIP_ADJACENCY_ARB, GL_TRIANGLE_STRIP_ADJACENCY_EXT.

The parameter type is a DrawElementsType.

The length of indices should be COMPSIZE(count,type).

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawElements.xhtml

glEdgeFlagPointer :: MonadIO m => GLsizei -> Ptr () -> m ()

Usage: glEdgeFlagPointer stride pointer

The length of pointer should be COMPSIZE(stride).

glEnableClientState :: MonadIO m => GLenum -> m ()

Usage: glEnableClientState array

The parameter array is a EnableCap, one of: GL_ALPHA_TEST, GL_ASYNC_DRAW_PIXELS_SGIX, GL_ASYNC_HISTOGRAM_SGIX, GL_ASYNC_READ_PIXELS_SGIX, GL_ASYNC_TEX_IMAGE_SGIX, GL_AUTO_NORMAL, GL_BLEND, GL_CALLIGRAPHIC_FRAGMENT_SGIX, GL_CLIP_PLANE0, GL_CLIP_PLANE1, GL_CLIP_PLANE2, GL_CLIP_PLANE3, GL_CLIP_PLANE4, GL_CLIP_PLANE5, GL_COLOR_ARRAY, GL_COLOR_LOGIC_OP, GL_COLOR_MATERIAL, GL_COLOR_TABLE_SGI, GL_CONVOLUTION_1D_EXT, GL_CONVOLUTION_2D_EXT, GL_CULL_FACE, GL_DEPTH_TEST, GL_DITHER, GL_EDGE_FLAG_ARRAY, GL_FOG, GL_FOG_OFFSET_SGIX, GL_FRAGMENT_COLOR_MATERIAL_SGIX, GL_FRAGMENT_LIGHT0_SGIX, GL_FRAGMENT_LIGHT1_SGIX, GL_FRAGMENT_LIGHT2_SGIX, GL_FRAGMENT_LIGHT3_SGIX, GL_FRAGMENT_LIGHT4_SGIX, GL_FRAGMENT_LIGHT5_SGIX, GL_FRAGMENT_LIGHT6_SGIX, GL_FRAGMENT_LIGHT7_SGIX, GL_FRAGMENT_LIGHTING_SGIX, GL_FRAMEZOOM_SGIX, GL_HISTOGRAM_EXT, GL_INDEX_ARRAY, GL_INDEX_LOGIC_OP, GL_INTERLACE_SGIX, GL_IR_INSTRUMENT1_SGIX, GL_LIGHT0, GL_LIGHT1, GL_LIGHT2, GL_LIGHT3, GL_LIGHT4, GL_LIGHT5, GL_LIGHT6, GL_LIGHT7, GL_LIGHTING, GL_LINE_SMOOTH, GL_LINE_STIPPLE, GL_MAP1_COLOR_4, GL_MAP1_INDEX, GL_MAP1_NORMAL, GL_MAP1_TEXTURE_COORD_1, GL_MAP1_TEXTURE_COORD_2, GL_MAP1_TEXTURE_COORD_3, GL_MAP1_TEXTURE_COORD_4, GL_MAP1_VERTEX_3, GL_MAP1_VERTEX_4, GL_MAP2_COLOR_4, GL_MAP2_INDEX, GL_MAP2_NORMAL, GL_MAP2_TEXTURE_COORD_1, GL_MAP2_TEXTURE_COORD_2, GL_MAP2_TEXTURE_COORD_3, GL_MAP2_TEXTURE_COORD_4, GL_MAP2_VERTEX_3, GL_MAP2_VERTEX_4, GL_MINMAX_EXT, GL_MULTISAMPLE_SGIS, GL_NORMALIZE, GL_NORMAL_ARRAY, GL_PIXEL_TEXTURE_SGIS, GL_PIXEL_TEX_GEN_SGIX, GL_POINT_SMOOTH, GL_POLYGON_OFFSET_FILL, GL_POLYGON_OFFSET_LINE, GL_POLYGON_OFFSET_POINT, GL_POLYGON_SMOOTH, GL_POLYGON_STIPPLE, GL_POST_COLOR_MATRIX_COLOR_TABLE_SGI, GL_POST_CONVOLUTION_COLOR_TABLE_SGI, GL_REFERENCE_PLANE_SGIX, GL_RESCALE_NORMAL_EXT, GL_SAMPLE_ALPHA_TO_MASK_SGIS, GL_SAMPLE_ALPHA_TO_ONE_SGIS, GL_SAMPLE_MASK_SGIS, GL_SCISSOR_TEST, GL_SEPARABLE_2D_EXT, GL_SHARED_TEXTURE_PALETTE_EXT, GL_SPRITE_SGIX, GL_STENCIL_TEST, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D_EXT, GL_TEXTURE_4D_SGIS, GL_TEXTURE_COLOR_TABLE_SGI, GL_TEXTURE_COORD_ARRAY, GL_TEXTURE_GEN_Q, GL_TEXTURE_GEN_R, GL_TEXTURE_GEN_S, GL_TEXTURE_GEN_T, GL_VERTEX_ARRAY.

glGenTextures :: MonadIO m => GLsizei -> Ptr GLuint -> m ()

Usage: glGenTextures n textures

The parameter textures is a Texture.

The length of textures should be n.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGenTextures.xhtml

glGetPointerv :: MonadIO m => GLenum -> Ptr () -> m ()

Usage: glGetPointerv pname params

The parameter pname is a GetPointervPName, one of: GL_COLOR_ARRAY_POINTER, GL_COLOR_ARRAY_POINTER_EXT, GL_EDGE_FLAG_ARRAY_POINTER, GL_EDGE_FLAG_ARRAY_POINTER_EXT, GL_FEEDBACK_BUFFER_POINTER, GL_INDEX_ARRAY_POINTER, GL_INDEX_ARRAY_POINTER_EXT, GL_INSTRUMENT_BUFFER_POINTER_SGIX, GL_NORMAL_ARRAY_POINTER, GL_NORMAL_ARRAY_POINTER_EXT, GL_SELECTION_BUFFER_POINTER, GL_TEXTURE_COORD_ARRAY_POINTER, GL_TEXTURE_COORD_ARRAY_POINTER_EXT, GL_VERTEX_ARRAY_POINTER, GL_VERTEX_ARRAY_POINTER_EXT.

The length of params should be 1.

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetPointerv.xhtml

glIndexPointer :: MonadIO m => GLenum -> GLsizei -> Ptr () -> m ()

Usage: glIndexPointer type stride pointer

The parameter type is a IndexPointerType, one of: GL_DOUBLE, GL_FLOAT, GL_INT, GL_SHORT.

The length of pointer should be COMPSIZE(type,stride).

glIndexub :: MonadIO m => GLubyte -> m ()

Usage: glIndexub c

The parameter c is a ColorIndexValueUB.

The vector equivalent of this command is glIndexubv.

glIndexubv :: MonadIO m => Ptr GLubyte -> m ()

Usage: glIndexubv c

The parameter c is a ColorIndexValueUB.

The length of c should be 1.

glInterleavedArrays :: MonadIO m => GLenum -> GLsizei -> Ptr () -> m ()

Usage: glInterleavedArrays format stride pointer

The parameter format is a InterleavedArrayFormat, one of: GL_C3F_V3F, GL_C4F_N3F_V3F, GL_C4UB_V2F, GL_C4UB_V3F, GL_N3F_V3F, GL_T2F_C3F_V3F, GL_T2F_C4F_N3F_V3F, GL_T2F_C4UB_V3F, GL_T2F_N3F_V3F, GL_T2F_V3F, GL_T4F_C4F_N3F_V4F, GL_T4F_V4F, GL_V2F, GL_V3F.

The length of pointer should be COMPSIZE(format,stride).

glIsTexture :: MonadIO m => GLuint -> m GLboolean

Usage: glIsTexture texture

The parameter texture is a Texture.

Manual page: https://www.opengl.org/sdk/docs/man/html/glIsTexture.xhtml

glNormalPointer :: MonadIO m => GLenum -> GLsizei -> Ptr () -> m ()

Usage: glNormalPointer type stride pointer

The parameter type is a NormalPointerType, one of: GL_BYTE, GL_DOUBLE, GL_FLOAT, GL_INT, GL_SHORT.

The length of pointer should be COMPSIZE(type,stride).

glPolygonOffset :: MonadIO m => GLfloat -> GLfloat -> m ()

Usage: glPolygonOffset factor units

Manual page: https://www.opengl.org/sdk/docs/man/html/glPolygonOffset.xhtml

glPopClientAttrib :: MonadIO m => m ()

Usage: glPopClientAttrib

glPrioritizeTextures :: MonadIO m => GLsizei -> Ptr GLuint -> Ptr GLfloat -> m ()

Usage: glPrioritizeTextures n textures priorities

The parameter textures is a Texture.

The length of textures should be n.

The length of priorities should be n.

glPushClientAttrib :: MonadIO m => GLbitfield -> m ()

Usage: glPushClientAttrib mask

The parameter mask is a ClientAttribMask, one of: GL_CLIENT_ALL_ATTRIB_BITS, GL_CLIENT_PIXEL_STORE_BIT, GL_CLIENT_VERTEX_ARRAY_BIT.

glTexCoordPointer :: MonadIO m => GLint -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glTexCoordPointer size type stride pointer

The parameter type is a TexCoordPointerType, one of: GL_DOUBLE, GL_FLOAT, GL_INT, GL_SHORT.

The length of pointer should be COMPSIZE(size,type,stride).

glTexSubImage1D :: MonadIO m => GLenum -> GLint -> GLint -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexSubImage1D target level xoffset width format type pixels

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, GL_UNSIGNED_BYTE_3_3_2_EXT, GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, GL_UNSIGNED_INT_10_10_10_2_EXT, GL_UNSIGNED_INT_8_8_8_8, GL_UNSIGNED_INT_8_8_8_8_EXT, GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, GL_UNSIGNED_SHORT_4_4_4_4_EXT, GL_UNSIGNED_SHORT_5_5_5_1, GL_UNSIGNED_SHORT_5_5_5_1_EXT.

The length of pixels should be COMPSIZE(format,type,width).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexSubImage1D.xhtml

glTexSubImage2D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexSubImage2D target level xoffset yoffset width height format type pixels

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, GL_UNSIGNED_BYTE_3_3_2_EXT, GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, GL_UNSIGNED_INT_10_10_10_2_EXT, GL_UNSIGNED_INT_8_8_8_8, GL_UNSIGNED_INT_8_8_8_8_EXT, GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, GL_UNSIGNED_SHORT_4_4_4_4_EXT, GL_UNSIGNED_SHORT_5_5_5_1, GL_UNSIGNED_SHORT_5_5_5_1_EXT.

The length of pixels should be COMPSIZE(format,type,width,height).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexSubImage2D.xhtml

glVertexPointer :: MonadIO m => GLint -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glVertexPointer size type stride pointer

The parameter type is a VertexPointerType, one of: GL_DOUBLE, GL_FLOAT, GL_INT, GL_SHORT.

The length of pointer should be COMPSIZE(size,type,stride).

pattern GL_2D :: (Num a, Eq a) => a

pattern GL_2_BYTES :: (Num a, Eq a) => a

pattern GL_3D :: (Num a, Eq a) => a

pattern GL_3D_COLOR :: (Num a, Eq a) => a

pattern GL_3D_COLOR_TEXTURE :: (Num a, Eq a) => a

pattern GL_3_BYTES :: (Num a, Eq a) => a

pattern GL_4D_COLOR_TEXTURE :: (Num a, Eq a) => a

pattern GL_4_BYTES :: (Num a, Eq a) => a

pattern GL_ACCUM :: (Num a, Eq a) => a

pattern GL_ACCUM_ALPHA_BITS :: (Num a, Eq a) => a

pattern GL_ACCUM_BLUE_BITS :: (Num a, Eq a) => a

pattern GL_ACCUM_BUFFER_BIT :: (Num a, Eq a) => a

pattern GL_ACCUM_CLEAR_VALUE :: (Num a, Eq a) => a

pattern GL_ACCUM_GREEN_BITS :: (Num a, Eq a) => a

pattern GL_ACCUM_RED_BITS :: (Num a, Eq a) => a

pattern GL_ADD :: (Num a, Eq a) => a

pattern GL_ALL_ATTRIB_BITS :: (Num a, Eq a) => a

pattern GL_ALPHA :: (Num a, Eq a) => a

pattern GL_ALPHA12 :: (Num a, Eq a) => a

pattern GL_ALPHA16 :: (Num a, Eq a) => a

pattern GL_ALPHA4 :: (Num a, Eq a) => a

pattern GL_ALPHA8 :: (Num a, Eq a) => a

pattern GL_ALPHA_BIAS :: (Num a, Eq a) => a

pattern GL_ALPHA_BITS :: (Num a, Eq a) => a

pattern GL_ALPHA_SCALE :: (Num a, Eq a) => a

pattern GL_ALPHA_TEST :: (Num a, Eq a) => a

pattern GL_ALPHA_TEST_FUNC :: (Num a, Eq a) => a

pattern GL_ALPHA_TEST_REF :: (Num a, Eq a) => a

pattern GL_ALWAYS :: (Num a, Eq a) => a

pattern GL_AMBIENT :: (Num a, Eq a) => a

pattern GL_AMBIENT_AND_DIFFUSE :: (Num a, Eq a) => a

pattern GL_AND :: (Num a, Eq a) => a

pattern GL_AND_INVERTED :: (Num a, Eq a) => a

pattern GL_AND_REVERSE :: (Num a, Eq a) => a

pattern GL_ATTRIB_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_AUTO_NORMAL :: (Num a, Eq a) => a

pattern GL_AUX0 :: (Num a, Eq a) => a

pattern GL_AUX1 :: (Num a, Eq a) => a

pattern GL_AUX2 :: (Num a, Eq a) => a

pattern GL_AUX3 :: (Num a, Eq a) => a

pattern GL_AUX_BUFFERS :: (Num a, Eq a) => a

pattern GL_BACK :: (Num a, Eq a) => a

pattern GL_BACK_LEFT :: (Num a, Eq a) => a

pattern GL_BACK_RIGHT :: (Num a, Eq a) => a

pattern GL_BITMAP :: (Num a, Eq a) => a

pattern GL_BITMAP_TOKEN :: (Num a, Eq a) => a

pattern GL_BLEND :: (Num a, Eq a) => a

pattern GL_BLEND_DST :: (Num a, Eq a) => a

pattern GL_BLEND_SRC :: (Num a, Eq a) => a

pattern GL_BLUE :: (Num a, Eq a) => a

pattern GL_BLUE_BIAS :: (Num a, Eq a) => a

pattern GL_BLUE_BITS :: (Num a, Eq a) => a

pattern GL_BLUE_SCALE :: (Num a, Eq a) => a

pattern GL_BYTE :: (Num a, Eq a) => a

pattern GL_C3F_V3F :: (Num a, Eq a) => a

pattern GL_C4F_N3F_V3F :: (Num a, Eq a) => a

pattern GL_C4UB_V2F :: (Num a, Eq a) => a

pattern GL_C4UB_V3F :: (Num a, Eq a) => a

pattern GL_CCW :: (Num a, Eq a) => a

pattern GL_CLAMP :: (Num a, Eq a) => a

pattern GL_CLEAR :: (Num a, Eq a) => a

pattern GL_CLIENT_ALL_ATTRIB_BITS :: (Num a, Eq a) => a

pattern GL_CLIENT_ATTRIB_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_CLIENT_PIXEL_STORE_BIT :: (Num a, Eq a) => a

pattern GL_CLIENT_VERTEX_ARRAY_BIT :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE0 :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE1 :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE2 :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE3 :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE4 :: (Num a, Eq a) => a

pattern GL_CLIP_PLANE5 :: (Num a, Eq a) => a

pattern GL_COEFF :: (Num a, Eq a) => a

pattern GL_COLOR :: (Num a, Eq a) => a

pattern GL_COLOR_ARRAY :: (Num a, Eq a) => a

pattern GL_COLOR_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_COLOR_ARRAY_SIZE :: (Num a, Eq a) => a

pattern GL_COLOR_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_COLOR_ARRAY_TYPE :: (Num a, Eq a) => a

pattern GL_COLOR_BUFFER_BIT :: (Num a, Eq a) => a

pattern GL_COLOR_CLEAR_VALUE :: (Num a, Eq a) => a

pattern GL_COLOR_INDEX :: (Num a, Eq a) => a

pattern GL_COLOR_INDEXES :: (Num a, Eq a) => a

pattern GL_COLOR_LOGIC_OP :: (Num a, Eq a) => a

pattern GL_COLOR_MATERIAL :: (Num a, Eq a) => a

pattern GL_COLOR_MATERIAL_FACE :: (Num a, Eq a) => a

pattern GL_COLOR_MATERIAL_PARAMETER :: (Num a, Eq a) => a

pattern GL_COLOR_WRITEMASK :: (Num a, Eq a) => a

pattern GL_COMPILE :: (Num a, Eq a) => a

pattern GL_COMPILE_AND_EXECUTE :: (Num a, Eq a) => a

pattern GL_CONSTANT_ATTENUATION :: (Num a, Eq a) => a

pattern GL_COPY :: (Num a, Eq a) => a

pattern GL_COPY_INVERTED :: (Num a, Eq a) => a

pattern GL_COPY_PIXEL_TOKEN :: (Num a, Eq a) => a

pattern GL_CULL_FACE :: (Num a, Eq a) => a

pattern GL_CULL_FACE_MODE :: (Num a, Eq a) => a

pattern GL_CURRENT_BIT :: (Num a, Eq a) => a

pattern GL_CURRENT_COLOR :: (Num a, Eq a) => a

pattern GL_CURRENT_INDEX :: (Num a, Eq a) => a

pattern GL_CURRENT_NORMAL :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_COLOR :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_DISTANCE :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_INDEX :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_POSITION :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_POSITION_VALID :: (Num a, Eq a) => a

pattern GL_CURRENT_RASTER_TEXTURE_COORDS :: (Num a, Eq a) => a

pattern GL_CURRENT_TEXTURE_COORDS :: (Num a, Eq a) => a

pattern GL_CW :: (Num a, Eq a) => a

pattern GL_DECAL :: (Num a, Eq a) => a

pattern GL_DECR :: (Num a, Eq a) => a

pattern GL_DEPTH :: (Num a, Eq a) => a

pattern GL_DEPTH_BIAS :: (Num a, Eq a) => a

pattern GL_DEPTH_BITS :: (Num a, Eq a) => a

pattern GL_DEPTH_BUFFER_BIT :: (Num a, Eq a) => a

pattern GL_DEPTH_CLEAR_VALUE :: (Num a, Eq a) => a

pattern GL_DEPTH_COMPONENT :: (Num a, Eq a) => a

pattern GL_DEPTH_FUNC :: (Num a, Eq a) => a

pattern GL_DEPTH_RANGE :: (Num a, Eq a) => a

pattern GL_DEPTH_SCALE :: (Num a, Eq a) => a

pattern GL_DEPTH_TEST :: (Num a, Eq a) => a

pattern GL_DEPTH_WRITEMASK :: (Num a, Eq a) => a

pattern GL_DIFFUSE :: (Num a, Eq a) => a

pattern GL_DITHER :: (Num a, Eq a) => a

pattern GL_DOMAIN :: (Num a, Eq a) => a

pattern GL_DONT_CARE :: (Num a, Eq a) => a

pattern GL_DOUBLE :: (Num a, Eq a) => a

pattern GL_DOUBLEBUFFER :: (Num a, Eq a) => a

pattern GL_DRAW_BUFFER :: (Num a, Eq a) => a

pattern GL_DRAW_PIXEL_TOKEN :: (Num a, Eq a) => a

pattern GL_DST_ALPHA :: (Num a, Eq a) => a

pattern GL_DST_COLOR :: (Num a, Eq a) => a

pattern GL_EDGE_FLAG :: (Num a, Eq a) => a

pattern GL_EDGE_FLAG_ARRAY :: (Num a, Eq a) => a

pattern GL_EDGE_FLAG_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_EDGE_FLAG_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_EMISSION :: (Num a, Eq a) => a

pattern GL_ENABLE_BIT :: (Num a, Eq a) => a

pattern GL_EQUAL :: (Num a, Eq a) => a

pattern GL_EQUIV :: (Num a, Eq a) => a

pattern GL_EVAL_BIT :: (Num a, Eq a) => a

pattern GL_EXP :: (Num a, Eq a) => a

pattern GL_EXP2 :: (Num a, Eq a) => a

pattern GL_EXTENSIONS :: (Num a, Eq a) => a

pattern GL_EYE_LINEAR :: (Num a, Eq a) => a

pattern GL_EYE_PLANE :: (Num a, Eq a) => a

pattern GL_FALSE :: (Num a, Eq a) => a

pattern GL_FASTEST :: (Num a, Eq a) => a

pattern GL_FEEDBACK :: (Num a, Eq a) => a

pattern GL_FEEDBACK_BUFFER_POINTER :: (Num a, Eq a) => a

pattern GL_FEEDBACK_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_FEEDBACK_BUFFER_TYPE :: (Num a, Eq a) => a

pattern GL_FILL :: (Num a, Eq a) => a

pattern GL_FLAT :: (Num a, Eq a) => a

pattern GL_FLOAT :: (Num a, Eq a) => a

pattern GL_FOG :: (Num a, Eq a) => a

pattern GL_FOG_BIT :: (Num a, Eq a) => a

pattern GL_FOG_COLOR :: (Num a, Eq a) => a

pattern GL_FOG_DENSITY :: (Num a, Eq a) => a

pattern GL_FOG_END :: (Num a, Eq a) => a

pattern GL_FOG_HINT :: (Num a, Eq a) => a

pattern GL_FOG_INDEX :: (Num a, Eq a) => a

pattern GL_FOG_MODE :: (Num a, Eq a) => a

pattern GL_FOG_START :: (Num a, Eq a) => a

pattern GL_FRONT :: (Num a, Eq a) => a

pattern GL_FRONT_AND_BACK :: (Num a, Eq a) => a

pattern GL_FRONT_FACE :: (Num a, Eq a) => a

pattern GL_FRONT_LEFT :: (Num a, Eq a) => a

pattern GL_FRONT_RIGHT :: (Num a, Eq a) => a

pattern GL_GEQUAL :: (Num a, Eq a) => a

pattern GL_GREATER :: (Num a, Eq a) => a

pattern GL_GREEN :: (Num a, Eq a) => a

pattern GL_GREEN_BIAS :: (Num a, Eq a) => a

pattern GL_GREEN_BITS :: (Num a, Eq a) => a

pattern GL_GREEN_SCALE :: (Num a, Eq a) => a

pattern GL_HINT_BIT :: (Num a, Eq a) => a

pattern GL_INCR :: (Num a, Eq a) => a

pattern GL_INDEX_ARRAY :: (Num a, Eq a) => a

pattern GL_INDEX_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_INDEX_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_INDEX_ARRAY_TYPE :: (Num a, Eq a) => a

pattern GL_INDEX_BITS :: (Num a, Eq a) => a

pattern GL_INDEX_CLEAR_VALUE :: (Num a, Eq a) => a

pattern GL_INDEX_LOGIC_OP :: (Num a, Eq a) => a

pattern GL_INDEX_MODE :: (Num a, Eq a) => a

pattern GL_INDEX_OFFSET :: (Num a, Eq a) => a

pattern GL_INDEX_SHIFT :: (Num a, Eq a) => a

pattern GL_INDEX_WRITEMASK :: (Num a, Eq a) => a

pattern GL_INT :: (Num a, Eq a) => a

pattern GL_INTENSITY :: (Num a, Eq a) => a

pattern GL_INTENSITY12 :: (Num a, Eq a) => a

pattern GL_INTENSITY16 :: (Num a, Eq a) => a

pattern GL_INTENSITY4 :: (Num a, Eq a) => a

pattern GL_INTENSITY8 :: (Num a, Eq a) => a

pattern GL_INVALID_ENUM :: (Num a, Eq a) => a

pattern GL_INVALID_OPERATION :: (Num a, Eq a) => a

pattern GL_INVALID_VALUE :: (Num a, Eq a) => a

pattern GL_INVERT :: (Num a, Eq a) => a

pattern GL_KEEP :: (Num a, Eq a) => a

pattern GL_LEFT :: (Num a, Eq a) => a

pattern GL_LEQUAL :: (Num a, Eq a) => a

pattern GL_LESS :: (Num a, Eq a) => a

pattern GL_LIGHT0 :: (Num a, Eq a) => a

pattern GL_LIGHT1 :: (Num a, Eq a) => a

pattern GL_LIGHT2 :: (Num a, Eq a) => a

pattern GL_LIGHT3 :: (Num a, Eq a) => a

pattern GL_LIGHT4 :: (Num a, Eq a) => a

pattern GL_LIGHT5 :: (Num a, Eq a) => a

pattern GL_LIGHT6 :: (Num a, Eq a) => a

pattern GL_LIGHT7 :: (Num a, Eq a) => a

pattern GL_LIGHTING :: (Num a, Eq a) => a

pattern GL_LIGHTING_BIT :: (Num a, Eq a) => a

pattern GL_LIGHT_MODEL_AMBIENT :: (Num a, Eq a) => a

pattern GL_LIGHT_MODEL_LOCAL_VIEWER :: (Num a, Eq a) => a

pattern GL_LIGHT_MODEL_TWO_SIDE :: (Num a, Eq a) => a

pattern GL_LINE :: (Num a, Eq a) => a

pattern GL_LINEAR :: (Num a, Eq a) => a

pattern GL_LINEAR_ATTENUATION :: (Num a, Eq a) => a

pattern GL_LINEAR_MIPMAP_LINEAR :: (Num a, Eq a) => a

pattern GL_LINEAR_MIPMAP_NEAREST :: (Num a, Eq a) => a

pattern GL_LINES :: (Num a, Eq a) => a

pattern GL_LINE_BIT :: (Num a, Eq a) => a

pattern GL_LINE_LOOP :: (Num a, Eq a) => a

pattern GL_LINE_RESET_TOKEN :: (Num a, Eq a) => a

pattern GL_LINE_SMOOTH :: (Num a, Eq a) => a

pattern GL_LINE_SMOOTH_HINT :: (Num a, Eq a) => a

pattern GL_LINE_STIPPLE :: (Num a, Eq a) => a

pattern GL_LINE_STIPPLE_PATTERN :: (Num a, Eq a) => a

pattern GL_LINE_STIPPLE_REPEAT :: (Num a, Eq a) => a

pattern GL_LINE_STRIP :: (Num a, Eq a) => a

pattern GL_LINE_TOKEN :: (Num a, Eq a) => a

pattern GL_LINE_WIDTH :: (Num a, Eq a) => a

pattern GL_LINE_WIDTH_GRANULARITY :: (Num a, Eq a) => a

pattern GL_LINE_WIDTH_RANGE :: (Num a, Eq a) => a

pattern GL_LIST_BASE :: (Num a, Eq a) => a

pattern GL_LIST_BIT :: (Num a, Eq a) => a

pattern GL_LIST_INDEX :: (Num a, Eq a) => a

pattern GL_LIST_MODE :: (Num a, Eq a) => a

pattern GL_LOAD :: (Num a, Eq a) => a

pattern GL_LOGIC_OP :: (Num a, Eq a) => a

pattern GL_LOGIC_OP_MODE :: (Num a, Eq a) => a

pattern GL_LUMINANCE :: (Num a, Eq a) => a

pattern GL_LUMINANCE12 :: (Num a, Eq a) => a

pattern GL_LUMINANCE12_ALPHA12 :: (Num a, Eq a) => a

pattern GL_LUMINANCE12_ALPHA4 :: (Num a, Eq a) => a

pattern GL_LUMINANCE16 :: (Num a, Eq a) => a

pattern GL_LUMINANCE16_ALPHA16 :: (Num a, Eq a) => a

pattern GL_LUMINANCE4 :: (Num a, Eq a) => a

pattern GL_LUMINANCE4_ALPHA4 :: (Num a, Eq a) => a

pattern GL_LUMINANCE6_ALPHA2 :: (Num a, Eq a) => a

pattern GL_LUMINANCE8 :: (Num a, Eq a) => a

pattern GL_LUMINANCE8_ALPHA8 :: (Num a, Eq a) => a

pattern GL_LUMINANCE_ALPHA :: (Num a, Eq a) => a

pattern GL_MAP1_COLOR_4 :: (Num a, Eq a) => a

pattern GL_MAP1_GRID_DOMAIN :: (Num a, Eq a) => a

pattern GL_MAP1_GRID_SEGMENTS :: (Num a, Eq a) => a

pattern GL_MAP1_INDEX :: (Num a, Eq a) => a

pattern GL_MAP1_NORMAL :: (Num a, Eq a) => a

pattern GL_MAP1_TEXTURE_COORD_1 :: (Num a, Eq a) => a

pattern GL_MAP1_TEXTURE_COORD_2 :: (Num a, Eq a) => a

pattern GL_MAP1_TEXTURE_COORD_3 :: (Num a, Eq a) => a

pattern GL_MAP1_TEXTURE_COORD_4 :: (Num a, Eq a) => a

pattern GL_MAP1_VERTEX_3 :: (Num a, Eq a) => a

pattern GL_MAP1_VERTEX_4 :: (Num a, Eq a) => a

pattern GL_MAP2_COLOR_4 :: (Num a, Eq a) => a

pattern GL_MAP2_GRID_DOMAIN :: (Num a, Eq a) => a

pattern GL_MAP2_GRID_SEGMENTS :: (Num a, Eq a) => a

pattern GL_MAP2_INDEX :: (Num a, Eq a) => a

pattern GL_MAP2_NORMAL :: (Num a, Eq a) => a

pattern GL_MAP2_TEXTURE_COORD_1 :: (Num a, Eq a) => a

pattern GL_MAP2_TEXTURE_COORD_2 :: (Num a, Eq a) => a

pattern GL_MAP2_TEXTURE_COORD_3 :: (Num a, Eq a) => a

pattern GL_MAP2_TEXTURE_COORD_4 :: (Num a, Eq a) => a

pattern GL_MAP2_VERTEX_3 :: (Num a, Eq a) => a

pattern GL_MAP2_VERTEX_4 :: (Num a, Eq a) => a

pattern GL_MAP_COLOR :: (Num a, Eq a) => a

pattern GL_MAP_STENCIL :: (Num a, Eq a) => a

pattern GL_MATRIX_MODE :: (Num a, Eq a) => a

pattern GL_MAX_ATTRIB_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_CLIENT_ATTRIB_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_CLIP_PLANES :: (Num a, Eq a) => a

pattern GL_MAX_EVAL_ORDER :: (Num a, Eq a) => a

pattern GL_MAX_LIGHTS :: (Num a, Eq a) => a

pattern GL_MAX_LIST_NESTING :: (Num a, Eq a) => a

pattern GL_MAX_MODELVIEW_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_NAME_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_PIXEL_MAP_TABLE :: (Num a, Eq a) => a

pattern GL_MAX_PROJECTION_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_TEXTURE_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_TEXTURE_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MAX_VIEWPORT_DIMS :: (Num a, Eq a) => a

pattern GL_MODELVIEW :: (Num a, Eq a) => a

pattern GL_MODELVIEW_MATRIX :: (Num a, Eq a) => a

pattern GL_MODELVIEW_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_MODULATE :: (Num a, Eq a) => a

pattern GL_MULT :: (Num a, Eq a) => a

pattern GL_N3F_V3F :: (Num a, Eq a) => a

pattern GL_NAME_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_NAND :: (Num a, Eq a) => a

pattern GL_NEAREST :: (Num a, Eq a) => a

pattern GL_NEAREST_MIPMAP_LINEAR :: (Num a, Eq a) => a

pattern GL_NEAREST_MIPMAP_NEAREST :: (Num a, Eq a) => a

pattern GL_NEVER :: (Num a, Eq a) => a

pattern GL_NICEST :: (Num a, Eq a) => a

pattern GL_NONE :: (Num a, Eq a) => a

pattern GL_NOOP :: (Num a, Eq a) => a

pattern GL_NOR :: (Num a, Eq a) => a

pattern GL_NORMALIZE :: (Num a, Eq a) => a

pattern GL_NORMAL_ARRAY :: (Num a, Eq a) => a

pattern GL_NORMAL_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_NORMAL_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_NORMAL_ARRAY_TYPE :: (Num a, Eq a) => a

pattern GL_NOTEQUAL :: (Num a, Eq a) => a

pattern GL_NO_ERROR :: (Num a, Eq a) => a

pattern GL_OBJECT_LINEAR :: (Num a, Eq a) => a

pattern GL_OBJECT_PLANE :: (Num a, Eq a) => a

pattern GL_ONE :: (Num a, Eq a) => a

pattern GL_ONE_MINUS_DST_ALPHA :: (Num a, Eq a) => a

pattern GL_ONE_MINUS_DST_COLOR :: (Num a, Eq a) => a

pattern GL_ONE_MINUS_SRC_ALPHA :: (Num a, Eq a) => a

pattern GL_ONE_MINUS_SRC_COLOR :: (Num a, Eq a) => a

pattern GL_OR :: (Num a, Eq a) => a

pattern GL_ORDER :: (Num a, Eq a) => a

pattern GL_OR_INVERTED :: (Num a, Eq a) => a

pattern GL_OR_REVERSE :: (Num a, Eq a) => a

pattern GL_OUT_OF_MEMORY :: (Num a, Eq a) => a

pattern GL_PACK_ALIGNMENT :: (Num a, Eq a) => a

pattern GL_PACK_LSB_FIRST :: (Num a, Eq a) => a

pattern GL_PACK_ROW_LENGTH :: (Num a, Eq a) => a

pattern GL_PACK_SKIP_PIXELS :: (Num a, Eq a) => a

pattern GL_PACK_SKIP_ROWS :: (Num a, Eq a) => a

pattern GL_PACK_SWAP_BYTES :: (Num a, Eq a) => a

pattern GL_PASS_THROUGH_TOKEN :: (Num a, Eq a) => a

pattern GL_PERSPECTIVE_CORRECTION_HINT :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_A_TO_A :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_A_TO_A_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_B_TO_B :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_B_TO_B_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_G_TO_G :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_G_TO_G_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_A :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_A_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_B :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_B_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_G :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_G_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_I :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_I_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_R :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_I_TO_R_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_R_TO_R :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_R_TO_R_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_S_TO_S :: (Num a, Eq a) => a

pattern GL_PIXEL_MAP_S_TO_S_SIZE :: (Num a, Eq a) => a

pattern GL_PIXEL_MODE_BIT :: (Num a, Eq a) => a

pattern GL_POINT :: (Num a, Eq a) => a

pattern GL_POINTS :: (Num a, Eq a) => a

pattern GL_POINT_BIT :: (Num a, Eq a) => a

pattern GL_POINT_SIZE :: (Num a, Eq a) => a

pattern GL_POINT_SIZE_GRANULARITY :: (Num a, Eq a) => a

pattern GL_POINT_SIZE_RANGE :: (Num a, Eq a) => a

pattern GL_POINT_SMOOTH :: (Num a, Eq a) => a

pattern GL_POINT_SMOOTH_HINT :: (Num a, Eq a) => a

pattern GL_POINT_TOKEN :: (Num a, Eq a) => a

pattern GL_POLYGON :: (Num a, Eq a) => a

pattern GL_POLYGON_BIT :: (Num a, Eq a) => a

pattern GL_POLYGON_MODE :: (Num a, Eq a) => a

pattern GL_POLYGON_OFFSET_FACTOR :: (Num a, Eq a) => a

pattern GL_POLYGON_OFFSET_FILL :: (Num a, Eq a) => a

pattern GL_POLYGON_OFFSET_LINE :: (Num a, Eq a) => a

pattern GL_POLYGON_OFFSET_POINT :: (Num a, Eq a) => a

pattern GL_POLYGON_OFFSET_UNITS :: (Num a, Eq a) => a

pattern GL_POLYGON_SMOOTH :: (Num a, Eq a) => a

pattern GL_POLYGON_SMOOTH_HINT :: (Num a, Eq a) => a

pattern GL_POLYGON_STIPPLE :: (Num a, Eq a) => a

pattern GL_POLYGON_STIPPLE_BIT :: (Num a, Eq a) => a

pattern GL_POLYGON_TOKEN :: (Num a, Eq a) => a

pattern GL_POSITION :: (Num a, Eq a) => a

pattern GL_PROJECTION :: (Num a, Eq a) => a

pattern GL_PROJECTION_MATRIX :: (Num a, Eq a) => a

pattern GL_PROJECTION_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_PROXY_TEXTURE_1D :: (Num a, Eq a) => a

pattern GL_PROXY_TEXTURE_2D :: (Num a, Eq a) => a

pattern GL_Q :: (Num a, Eq a) => a

pattern GL_QUADRATIC_ATTENUATION :: (Num a, Eq a) => a

pattern GL_QUADS :: (Num a, Eq a) => a

pattern GL_QUAD_STRIP :: (Num a, Eq a) => a

pattern GL_R :: (Num a, Eq a) => a

pattern GL_R3_G3_B2 :: (Num a, Eq a) => a

pattern GL_READ_BUFFER :: (Num a, Eq a) => a

pattern GL_RED :: (Num a, Eq a) => a

pattern GL_RED_BIAS :: (Num a, Eq a) => a

pattern GL_RED_BITS :: (Num a, Eq a) => a

pattern GL_RED_SCALE :: (Num a, Eq a) => a

pattern GL_RENDER :: (Num a, Eq a) => a

pattern GL_RENDERER :: (Num a, Eq a) => a

pattern GL_RENDER_MODE :: (Num a, Eq a) => a

pattern GL_REPEAT :: (Num a, Eq a) => a

pattern GL_REPLACE :: (Num a, Eq a) => a

pattern GL_RETURN :: (Num a, Eq a) => a

pattern GL_RGB :: (Num a, Eq a) => a

pattern GL_RGB10 :: (Num a, Eq a) => a

pattern GL_RGB10_A2 :: (Num a, Eq a) => a

pattern GL_RGB12 :: (Num a, Eq a) => a

pattern GL_RGB16 :: (Num a, Eq a) => a

pattern GL_RGB4 :: (Num a, Eq a) => a

pattern GL_RGB5 :: (Num a, Eq a) => a

pattern GL_RGB5_A1 :: (Num a, Eq a) => a

pattern GL_RGB8 :: (Num a, Eq a) => a

pattern GL_RGBA :: (Num a, Eq a) => a

pattern GL_RGBA12 :: (Num a, Eq a) => a

pattern GL_RGBA16 :: (Num a, Eq a) => a

pattern GL_RGBA2 :: (Num a, Eq a) => a

pattern GL_RGBA4 :: (Num a, Eq a) => a

pattern GL_RGBA8 :: (Num a, Eq a) => a

pattern GL_RGBA_MODE :: (Num a, Eq a) => a

pattern GL_RIGHT :: (Num a, Eq a) => a

pattern GL_S :: (Num a, Eq a) => a

pattern GL_SCISSOR_BIT :: (Num a, Eq a) => a

pattern GL_SCISSOR_BOX :: (Num a, Eq a) => a

pattern GL_SCISSOR_TEST :: (Num a, Eq a) => a

pattern GL_SELECT :: (Num a, Eq a) => a

pattern GL_SELECTION_BUFFER_POINTER :: (Num a, Eq a) => a

pattern GL_SELECTION_BUFFER_SIZE :: (Num a, Eq a) => a

pattern GL_SET :: (Num a, Eq a) => a

pattern GL_SHADE_MODEL :: (Num a, Eq a) => a

pattern GL_SHININESS :: (Num a, Eq a) => a

pattern GL_SHORT :: (Num a, Eq a) => a

pattern GL_SMOOTH :: (Num a, Eq a) => a

pattern GL_SPECULAR :: (Num a, Eq a) => a

pattern GL_SPHERE_MAP :: (Num a, Eq a) => a

pattern GL_SPOT_CUTOFF :: (Num a, Eq a) => a

pattern GL_SPOT_DIRECTION :: (Num a, Eq a) => a

pattern GL_SPOT_EXPONENT :: (Num a, Eq a) => a

pattern GL_SRC_ALPHA :: (Num a, Eq a) => a

pattern GL_SRC_ALPHA_SATURATE :: (Num a, Eq a) => a

pattern GL_SRC_COLOR :: (Num a, Eq a) => a

pattern GL_STACK_OVERFLOW :: (Num a, Eq a) => a

pattern GL_STACK_UNDERFLOW :: (Num a, Eq a) => a

pattern GL_STENCIL :: (Num a, Eq a) => a

pattern GL_STENCIL_BITS :: (Num a, Eq a) => a

pattern GL_STENCIL_BUFFER_BIT :: (Num a, Eq a) => a

pattern GL_STENCIL_CLEAR_VALUE :: (Num a, Eq a) => a

pattern GL_STENCIL_FAIL :: (Num a, Eq a) => a

pattern GL_STENCIL_FUNC :: (Num a, Eq a) => a

pattern GL_STENCIL_INDEX :: (Num a, Eq a) => a

pattern GL_STENCIL_PASS_DEPTH_FAIL :: (Num a, Eq a) => a

pattern GL_STENCIL_PASS_DEPTH_PASS :: (Num a, Eq a) => a

pattern GL_STENCIL_REF :: (Num a, Eq a) => a

pattern GL_STENCIL_TEST :: (Num a, Eq a) => a

pattern GL_STENCIL_VALUE_MASK :: (Num a, Eq a) => a

pattern GL_STENCIL_WRITEMASK :: (Num a, Eq a) => a

pattern GL_STEREO :: (Num a, Eq a) => a

pattern GL_SUBPIXEL_BITS :: (Num a, Eq a) => a

pattern GL_T :: (Num a, Eq a) => a

pattern GL_T2F_C3F_V3F :: (Num a, Eq a) => a

pattern GL_T2F_C4F_N3F_V3F :: (Num a, Eq a) => a

pattern GL_T2F_C4UB_V3F :: (Num a, Eq a) => a

pattern GL_T2F_N3F_V3F :: (Num a, Eq a) => a

pattern GL_T2F_V3F :: (Num a, Eq a) => a

pattern GL_T4F_C4F_N3F_V4F :: (Num a, Eq a) => a

pattern GL_T4F_V4F :: (Num a, Eq a) => a

pattern GL_TEXTURE :: (Num a, Eq a) => a

pattern GL_TEXTURE_1D :: (Num a, Eq a) => a

pattern GL_TEXTURE_2D :: (Num a, Eq a) => a

pattern GL_TEXTURE_ALPHA_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_1D :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_2D :: (Num a, Eq a) => a

pattern GL_TEXTURE_BIT :: (Num a, Eq a) => a

pattern GL_TEXTURE_BLUE_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_BORDER :: (Num a, Eq a) => a

pattern GL_TEXTURE_BORDER_COLOR :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPONENTS :: (Num a, Eq a) => a

pattern GL_TEXTURE_COORD_ARRAY :: (Num a, Eq a) => a

pattern GL_TEXTURE_COORD_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_TEXTURE_COORD_ARRAY_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_COORD_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_TEXTURE_COORD_ARRAY_TYPE :: (Num a, Eq a) => a

pattern GL_TEXTURE_ENV :: (Num a, Eq a) => a

pattern GL_TEXTURE_ENV_COLOR :: (Num a, Eq a) => a

pattern GL_TEXTURE_ENV_MODE :: (Num a, Eq a) => a

pattern GL_TEXTURE_GEN_MODE :: (Num a, Eq a) => a

pattern GL_TEXTURE_GEN_Q :: (Num a, Eq a) => a

pattern GL_TEXTURE_GEN_R :: (Num a, Eq a) => a

pattern GL_TEXTURE_GEN_S :: (Num a, Eq a) => a

pattern GL_TEXTURE_GEN_T :: (Num a, Eq a) => a

pattern GL_TEXTURE_GREEN_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_HEIGHT :: (Num a, Eq a) => a

pattern GL_TEXTURE_INTENSITY_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_INTERNAL_FORMAT :: (Num a, Eq a) => a

pattern GL_TEXTURE_LUMINANCE_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_MAG_FILTER :: (Num a, Eq a) => a

pattern GL_TEXTURE_MATRIX :: (Num a, Eq a) => a

pattern GL_TEXTURE_MIN_FILTER :: (Num a, Eq a) => a

pattern GL_TEXTURE_PRIORITY :: (Num a, Eq a) => a

pattern GL_TEXTURE_RED_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_RESIDENT :: (Num a, Eq a) => a

pattern GL_TEXTURE_STACK_DEPTH :: (Num a, Eq a) => a

pattern GL_TEXTURE_WIDTH :: (Num a, Eq a) => a

pattern GL_TEXTURE_WRAP_S :: (Num a, Eq a) => a

pattern GL_TEXTURE_WRAP_T :: (Num a, Eq a) => a

pattern GL_TRANSFORM_BIT :: (Num a, Eq a) => a

pattern GL_TRIANGLES :: (Num a, Eq a) => a

pattern GL_TRIANGLE_FAN :: (Num a, Eq a) => a

pattern GL_TRIANGLE_STRIP :: (Num a, Eq a) => a

pattern GL_TRUE :: (Num a, Eq a) => a

pattern GL_UNPACK_ALIGNMENT :: (Num a, Eq a) => a

pattern GL_UNPACK_LSB_FIRST :: (Num a, Eq a) => a

pattern GL_UNPACK_ROW_LENGTH :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_PIXELS :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_ROWS :: (Num a, Eq a) => a

pattern GL_UNPACK_SWAP_BYTES :: (Num a, Eq a) => a

pattern GL_UNSIGNED_BYTE :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT :: (Num a, Eq a) => a

pattern GL_V2F :: (Num a, Eq a) => a

pattern GL_V3F :: (Num a, Eq a) => a

pattern GL_VENDOR :: (Num a, Eq a) => a

pattern GL_VERSION :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY_POINTER :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY_SIZE :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY_STRIDE :: (Num a, Eq a) => a

pattern GL_VERTEX_ARRAY_TYPE :: (Num a, Eq a) => a

pattern GL_VIEWPORT :: (Num a, Eq a) => a

pattern GL_VIEWPORT_BIT :: (Num a, Eq a) => a

pattern GL_XOR :: (Num a, Eq a) => a

pattern GL_ZERO :: (Num a, Eq a) => a

pattern GL_ZOOM_X :: (Num a, Eq a) => a

pattern GL_ZOOM_Y :: (Num a, Eq a) => a