Graphics.GL.Raw.Profile.Standard12

Graphics.GL.Raw.Profile.Standard12

module Graphics.GL.Raw.Profile.Standard11

glCopyTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage3D target level xoffset yoffset zoffset x y width height

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter x is a WinCoord.

The parameter y is a WinCoord.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCopyTexSubImage3D.xhtml

glDrawRangeElements :: MonadIO m => GLenum -> GLuint -> GLuint -> GLsizei -> GLenum -> Ptr () -> m ()

Usage: glDrawRangeElements mode start end count type indices

The parameter mode is a PrimitiveType, one of: GL_LINES, GL_LINES_ADJACENCY, GL_LINES_ADJACENCY_ARB, GL_LINES_ADJACENCY_EXT, GL_LINE_LOOP, GL_LINE_STRIP, GL_LINE_STRIP_ADJACENCY, GL_LINE_STRIP_ADJACENCY_ARB, GL_LINE_STRIP_ADJACENCY_EXT, GL_PATCHES, GL_PATCHES_EXT, GL_POINTS, GL_POLYGON, GL_QUADS, GL_QUADS_EXT, GL_QUAD_STRIP, GL_TRIANGLES, GL_TRIANGLES_ADJACENCY, GL_TRIANGLES_ADJACENCY_ARB, GL_TRIANGLES_ADJACENCY_EXT, GL_TRIANGLE_FAN, GL_TRIANGLE_STRIP, GL_TRIANGLE_STRIP_ADJACENCY, GL_TRIANGLE_STRIP_ADJACENCY_ARB, GL_TRIANGLE_STRIP_ADJACENCY_EXT.

The parameter type is a DrawElementsType.

The length of indices should be COMPSIZE(count,type).

Manual page: https://www.opengl.org/sdk/docs/man/html/glDrawRangeElements.xhtml

glTexImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexImage3D target level internalformat width height depth border format type pixels

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a TextureComponentCount.

The parameter border is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, GL_UNSIGNED_BYTE_3_3_2_EXT, GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, GL_UNSIGNED_INT_10_10_10_2_EXT, GL_UNSIGNED_INT_8_8_8_8, GL_UNSIGNED_INT_8_8_8_8_EXT, GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, GL_UNSIGNED_SHORT_4_4_4_4_EXT, GL_UNSIGNED_SHORT_5_5_5_1, GL_UNSIGNED_SHORT_5_5_5_1_EXT.

The length of pixels should be COMPSIZE(format,type,width,height,depth).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexImage3D.xhtml

glTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLenum -> Ptr () -> m ()

Usage: glTexSubImage3D target level xoffset yoffset zoffset width height depth format type pixels

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter type is a PixelType, one of: GL_BITMAP, GL_BYTE, GL_FLOAT, GL_INT, GL_SHORT, GL_UNSIGNED_BYTE, GL_UNSIGNED_BYTE_3_3_2, GL_UNSIGNED_BYTE_3_3_2_EXT, GL_UNSIGNED_INT, GL_UNSIGNED_INT_10_10_10_2, GL_UNSIGNED_INT_10_10_10_2_EXT, GL_UNSIGNED_INT_8_8_8_8, GL_UNSIGNED_INT_8_8_8_8_EXT, GL_UNSIGNED_SHORT, GL_UNSIGNED_SHORT_4_4_4_4, GL_UNSIGNED_SHORT_4_4_4_4_EXT, GL_UNSIGNED_SHORT_5_5_5_1, GL_UNSIGNED_SHORT_5_5_5_1_EXT.

The length of pixels should be COMPSIZE(format,type,width,height,depth).

Manual page: https://www.opengl.org/sdk/docs/man/html/glTexSubImage3D.xhtml

pattern GL_ALIASED_LINE_WIDTH_RANGE :: (Num a, Eq a) => a

pattern GL_ALIASED_POINT_SIZE_RANGE :: (Num a, Eq a) => a

pattern GL_BGR :: (Num a, Eq a) => a

pattern GL_BGRA :: (Num a, Eq a) => a

pattern GL_CLAMP_TO_EDGE :: (Num a, Eq a) => a

pattern GL_LIGHT_MODEL_COLOR_CONTROL :: (Num a, Eq a) => a

pattern GL_MAX_3D_TEXTURE_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENTS_INDICES :: (Num a, Eq a) => a

pattern GL_MAX_ELEMENTS_VERTICES :: (Num a, Eq a) => a

pattern GL_PACK_IMAGE_HEIGHT :: (Num a, Eq a) => a

pattern GL_PACK_SKIP_IMAGES :: (Num a, Eq a) => a

pattern GL_PROXY_TEXTURE_3D :: (Num a, Eq a) => a

pattern GL_RESCALE_NORMAL :: (Num a, Eq a) => a

pattern GL_SEPARATE_SPECULAR_COLOR :: (Num a, Eq a) => a

pattern GL_SINGLE_COLOR :: (Num a, Eq a) => a

pattern GL_SMOOTH_LINE_WIDTH_GRANULARITY :: (Num a, Eq a) => a

pattern GL_SMOOTH_LINE_WIDTH_RANGE :: (Num a, Eq a) => a

pattern GL_SMOOTH_POINT_SIZE_GRANULARITY :: (Num a, Eq a) => a

pattern GL_SMOOTH_POINT_SIZE_RANGE :: (Num a, Eq a) => a

pattern GL_TEXTURE_3D :: (Num a, Eq a) => a

pattern GL_TEXTURE_BASE_LEVEL :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_3D :: (Num a, Eq a) => a

pattern GL_TEXTURE_DEPTH :: (Num a, Eq a) => a

pattern GL_TEXTURE_MAX_LEVEL :: (Num a, Eq a) => a

pattern GL_TEXTURE_MAX_LOD :: (Num a, Eq a) => a

pattern GL_TEXTURE_MIN_LOD :: (Num a, Eq a) => a

pattern GL_TEXTURE_WRAP_R :: (Num a, Eq a) => a

pattern GL_UNPACK_IMAGE_HEIGHT :: (Num a, Eq a) => a

pattern GL_UNPACK_SKIP_IMAGES :: (Num a, Eq a) => a

pattern GL_UNSIGNED_BYTE_2_3_3_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_BYTE_3_3_2 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_10_10_10_2 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_2_10_10_10_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_8_8_8_8 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_INT_8_8_8_8_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_1_5_5_5_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_4_4_4_4 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_4_4_4_4_REV :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_5_5_5_1 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_5_6_5 :: (Num a, Eq a) => a

pattern GL_UNSIGNED_SHORT_5_6_5_REV :: (Num a, Eq a) => a