Graphics.GL.Raw.Profile.Standard13

Graphics.GL.Raw.Profile.Standard13

module Graphics.GL.Raw.Profile.Standard12

glActiveTexture :: MonadIO m => GLenum -> m ()

Usage: glActiveTexture texture

The parameter texture is a TextureUnit.

Manual page: https://www.opengl.org/sdk/docs/man/html/glActiveTexture.xhtml

glClientActiveTexture :: MonadIO m => GLenum -> m ()

Usage: glClientActiveTexture texture

The parameter texture is a TextureUnit.

glCompressedTexImage1D :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage1D target level internalformat width border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexImage1D.xhtml

glCompressedTexImage2D :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage2D target level internalformat width height border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexImage2D.xhtml

glCompressedTexImage3D :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage3D target level internalformat width height depth border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexImage3D.xhtml

glCompressedTexSubImage1D :: MonadIO m => GLenum -> GLint -> GLint -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage1D target level xoffset width format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexSubImage1D.xhtml

glCompressedTexSubImage2D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage2D target level xoffset yoffset width height format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexSubImage2D.xhtml

glCompressedTexSubImage3D :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage3D target level xoffset yoffset zoffset width height depth format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

Manual page: https://www.opengl.org/sdk/docs/man/html/glCompressedTexSubImage3D.xhtml

glGetCompressedTexImage :: MonadIO m => GLenum -> GLint -> Ptr () -> m ()

Usage: glGetCompressedTexImage target level img

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter img is a CompressedTextureARB.

The length of img should be COMPSIZE(target,level).

Manual page: https://www.opengl.org/sdk/docs/man/html/glGetCompressedTexImage.xhtml

glLoadTransposeMatrixd :: MonadIO m => Ptr GLdouble -> m ()

Usage: glLoadTransposeMatrixd m

The length of m should be 16.

glLoadTransposeMatrixf :: MonadIO m => Ptr GLfloat -> m ()

Usage: glLoadTransposeMatrixf m

The length of m should be 16.

glMultTransposeMatrixd :: MonadIO m => Ptr GLdouble -> m ()

Usage: glMultTransposeMatrixd m

The length of m should be 16.

glMultTransposeMatrixf :: MonadIO m => Ptr GLfloat -> m ()

Usage: glMultTransposeMatrixf m

The length of m should be 16.

glMultiTexCoord1d :: MonadIO m => GLenum -> GLdouble -> m ()

Usage: glMultiTexCoord1d target s

The parameter target is a TextureUnit.

The parameter s is a CoordD.

The vector equivalent of this command is glMultiTexCoord1dv.

glMultiTexCoord1dv :: MonadIO m => GLenum -> Ptr GLdouble -> m ()

Usage: glMultiTexCoord1dv target v

The parameter target is a TextureUnit.

The parameter v is a CoordD.

The length of v should be 1.

glMultiTexCoord1f :: MonadIO m => GLenum -> GLfloat -> m ()

Usage: glMultiTexCoord1f target s

The parameter target is a TextureUnit.

The parameter s is a CoordF.

The vector equivalent of this command is glMultiTexCoord1fv.

glMultiTexCoord1fv :: MonadIO m => GLenum -> Ptr GLfloat -> m ()

Usage: glMultiTexCoord1fv target v

The parameter target is a TextureUnit.

The parameter v is a CoordF.

The length of v should be 1.

glMultiTexCoord1i :: MonadIO m => GLenum -> GLint -> m ()

Usage: glMultiTexCoord1i target s

The parameter target is a TextureUnit.

The parameter s is a CoordI.

The vector equivalent of this command is glMultiTexCoord1iv.

glMultiTexCoord1iv :: MonadIO m => GLenum -> Ptr GLint -> m ()

Usage: glMultiTexCoord1iv target v

The parameter target is a TextureUnit.

The parameter v is a CoordI.

The length of v should be 1.

glMultiTexCoord1s :: MonadIO m => GLenum -> GLshort -> m ()

Usage: glMultiTexCoord1s target s

The parameter target is a TextureUnit.

The parameter s is a CoordS.

The vector equivalent of this command is glMultiTexCoord1sv.

glMultiTexCoord1sv :: MonadIO m => GLenum -> Ptr GLshort -> m ()

Usage: glMultiTexCoord1sv target v

The parameter target is a TextureUnit.

The parameter v is a CoordS.

The length of v should be 1.

glMultiTexCoord2d :: MonadIO m => GLenum -> GLdouble -> GLdouble -> m ()

Usage: glMultiTexCoord2d target s t

The parameter target is a TextureUnit.

The parameter s is a CoordD.

The parameter t is a CoordD.

The vector equivalent of this command is glMultiTexCoord2dv.

glMultiTexCoord2dv :: MonadIO m => GLenum -> Ptr GLdouble -> m ()

Usage: glMultiTexCoord2dv target v

The parameter target is a TextureUnit.

The parameter v is a CoordD.

The length of v should be 2.

glMultiTexCoord2f :: MonadIO m => GLenum -> GLfloat -> GLfloat -> m ()

Usage: glMultiTexCoord2f target s t

The parameter target is a TextureUnit.

The parameter s is a CoordF.

The parameter t is a CoordF.

The vector equivalent of this command is glMultiTexCoord2fv.

glMultiTexCoord2fv :: MonadIO m => GLenum -> Ptr GLfloat -> m ()

Usage: glMultiTexCoord2fv target v

The parameter target is a TextureUnit.

The parameter v is a CoordF.

The length of v should be 2.

glMultiTexCoord2i :: MonadIO m => GLenum -> GLint -> GLint -> m ()

Usage: glMultiTexCoord2i target s t

The parameter target is a TextureUnit.

The parameter s is a CoordI.

The parameter t is a CoordI.

The vector equivalent of this command is glMultiTexCoord2iv.

glMultiTexCoord2iv :: MonadIO m => GLenum -> Ptr GLint -> m ()

Usage: glMultiTexCoord2iv target v

The parameter target is a TextureUnit.

The parameter v is a CoordI.

The length of v should be 2.

glMultiTexCoord2s :: MonadIO m => GLenum -> GLshort -> GLshort -> m ()

Usage: glMultiTexCoord2s target s t

The parameter target is a TextureUnit.

The parameter s is a CoordS.

The parameter t is a CoordS.

The vector equivalent of this command is glMultiTexCoord2sv.

glMultiTexCoord2sv :: MonadIO m => GLenum -> Ptr GLshort -> m ()

Usage: glMultiTexCoord2sv target v

The parameter target is a TextureUnit.

The parameter v is a CoordS.

The length of v should be 2.

glMultiTexCoord3d :: MonadIO m => GLenum -> GLdouble -> GLdouble -> GLdouble -> m ()

Usage: glMultiTexCoord3d target s t r

The parameter target is a TextureUnit.

The parameter s is a CoordD.

The parameter t is a CoordD.

The parameter r is a CoordD.

The vector equivalent of this command is glMultiTexCoord3dv.

glMultiTexCoord3dv :: MonadIO m => GLenum -> Ptr GLdouble -> m ()

Usage: glMultiTexCoord3dv target v

The parameter target is a TextureUnit.

The parameter v is a CoordD.

The length of v should be 3.

glMultiTexCoord3f :: MonadIO m => GLenum -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glMultiTexCoord3f target s t r

The parameter target is a TextureUnit.

The parameter s is a CoordF.

The parameter t is a CoordF.

The parameter r is a CoordF.

The vector equivalent of this command is glMultiTexCoord3fv.

glMultiTexCoord3fv :: MonadIO m => GLenum -> Ptr GLfloat -> m ()

Usage: glMultiTexCoord3fv target v

The parameter target is a TextureUnit.

The parameter v is a CoordF.

The length of v should be 3.

glMultiTexCoord3i :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> m ()

Usage: glMultiTexCoord3i target s t r

The parameter target is a TextureUnit.

The parameter s is a CoordI.

The parameter t is a CoordI.

The parameter r is a CoordI.

The vector equivalent of this command is glMultiTexCoord3iv.

glMultiTexCoord3iv :: MonadIO m => GLenum -> Ptr GLint -> m ()

Usage: glMultiTexCoord3iv target v

The parameter target is a TextureUnit.

The parameter v is a CoordI.

The length of v should be 3.

glMultiTexCoord3s :: MonadIO m => GLenum -> GLshort -> GLshort -> GLshort -> m ()

Usage: glMultiTexCoord3s target s t r

The parameter target is a TextureUnit.

The parameter s is a CoordS.

The parameter t is a CoordS.

The parameter r is a CoordS.

The vector equivalent of this command is glMultiTexCoord3sv.

glMultiTexCoord3sv :: MonadIO m => GLenum -> Ptr GLshort -> m ()

Usage: glMultiTexCoord3sv target v

The parameter target is a TextureUnit.

The parameter v is a CoordS.

The length of v should be 3.

glMultiTexCoord4d :: MonadIO m => GLenum -> GLdouble -> GLdouble -> GLdouble -> GLdouble -> m ()

Usage: glMultiTexCoord4d target s t r q

The parameter target is a TextureUnit.

The parameter s is a CoordD.

The parameter t is a CoordD.

The parameter r is a CoordD.

The parameter q is a CoordD.

The vector equivalent of this command is glMultiTexCoord4dv.

glMultiTexCoord4dv :: MonadIO m => GLenum -> Ptr GLdouble -> m ()

Usage: glMultiTexCoord4dv target v

The parameter target is a TextureUnit.

The parameter v is a CoordD.

The length of v should be 4.

glMultiTexCoord4f :: MonadIO m => GLenum -> GLfloat -> GLfloat -> GLfloat -> GLfloat -> m ()

Usage: glMultiTexCoord4f target s t r q

The parameter target is a TextureUnit.

The parameter s is a CoordF.

The parameter t is a CoordF.

The parameter r is a CoordF.

The parameter q is a CoordF.

The vector equivalent of this command is glMultiTexCoord4fv.

glMultiTexCoord4fv :: MonadIO m => GLenum -> Ptr GLfloat -> m ()

Usage: glMultiTexCoord4fv target v

The parameter target is a TextureUnit.

The parameter v is a CoordF.

The length of v should be 4.

glMultiTexCoord4i :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> m ()

Usage: glMultiTexCoord4i target s t r q

The parameter target is a TextureUnit.

The parameter s is a CoordI.

The parameter t is a CoordI.

The parameter r is a CoordI.

The parameter q is a CoordI.

The vector equivalent of this command is glMultiTexCoord4iv.

glMultiTexCoord4iv :: MonadIO m => GLenum -> Ptr GLint -> m ()

Usage: glMultiTexCoord4iv target v

The parameter target is a TextureUnit.

The parameter v is a CoordI.

The length of v should be 4.

glMultiTexCoord4s :: MonadIO m => GLenum -> GLshort -> GLshort -> GLshort -> GLshort -> m ()

Usage: glMultiTexCoord4s target s t r q

The parameter target is a TextureUnit.

The parameter s is a CoordS.

The parameter t is a CoordS.

The parameter r is a CoordS.

The parameter q is a CoordS.

The vector equivalent of this command is glMultiTexCoord4sv.

glMultiTexCoord4sv :: MonadIO m => GLenum -> Ptr GLshort -> m ()

Usage: glMultiTexCoord4sv target v

The parameter target is a TextureUnit.

The parameter v is a CoordS.

The length of v should be 4.

glSampleCoverage :: MonadIO m => GLfloat -> GLboolean -> m ()

Usage: glSampleCoverage value invert

The parameter invert is a Boolean, one of: GL_FALSE, GL_TRUE.

Manual page: https://www.opengl.org/sdk/docs/man/html/glSampleCoverage.xhtml

pattern GL_ACTIVE_TEXTURE :: (Num a, Eq a) => a

pattern GL_ADD_SIGNED :: (Num a, Eq a) => a

pattern GL_CLAMP_TO_BORDER :: (Num a, Eq a) => a

pattern GL_CLIENT_ACTIVE_TEXTURE :: (Num a, Eq a) => a

pattern GL_COMBINE :: (Num a, Eq a) => a

pattern GL_COMBINE_ALPHA :: (Num a, Eq a) => a

pattern GL_COMBINE_RGB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_ALPHA :: (Num a, Eq a) => a

pattern GL_COMPRESSED_INTENSITY :: (Num a, Eq a) => a

pattern GL_COMPRESSED_LUMINANCE :: (Num a, Eq a) => a

pattern GL_COMPRESSED_LUMINANCE_ALPHA :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGBA :: (Num a, Eq a) => a

pattern GL_COMPRESSED_TEXTURE_FORMATS :: (Num a, Eq a) => a

pattern GL_CONSTANT :: (Num a, Eq a) => a

pattern GL_DOT3_RGB :: (Num a, Eq a) => a

pattern GL_DOT3_RGBA :: (Num a, Eq a) => a

pattern GL_INTERPOLATE :: (Num a, Eq a) => a

pattern GL_MAX_CUBE_MAP_TEXTURE_SIZE :: (Num a, Eq a) => a

pattern GL_MAX_TEXTURE_UNITS :: (Num a, Eq a) => a

pattern GL_MULTISAMPLE :: (Num a, Eq a) => a

pattern GL_MULTISAMPLE_BIT :: (Num a, Eq a) => a

pattern GL_NORMAL_MAP :: (Num a, Eq a) => a

pattern GL_NUM_COMPRESSED_TEXTURE_FORMATS :: (Num a, Eq a) => a

pattern GL_OPERAND0_ALPHA :: (Num a, Eq a) => a

pattern GL_OPERAND0_RGB :: (Num a, Eq a) => a

pattern GL_OPERAND1_ALPHA :: (Num a, Eq a) => a

pattern GL_OPERAND1_RGB :: (Num a, Eq a) => a

pattern GL_OPERAND2_ALPHA :: (Num a, Eq a) => a

pattern GL_OPERAND2_RGB :: (Num a, Eq a) => a

pattern GL_PREVIOUS :: (Num a, Eq a) => a

pattern GL_PRIMARY_COLOR :: (Num a, Eq a) => a

pattern GL_PROXY_TEXTURE_CUBE_MAP :: (Num a, Eq a) => a

pattern GL_REFLECTION_MAP :: (Num a, Eq a) => a

pattern GL_RGB_SCALE :: (Num a, Eq a) => a

pattern GL_SAMPLES :: (Num a, Eq a) => a

pattern GL_SAMPLE_ALPHA_TO_COVERAGE :: (Num a, Eq a) => a

pattern GL_SAMPLE_ALPHA_TO_ONE :: (Num a, Eq a) => a

pattern GL_SAMPLE_BUFFERS :: (Num a, Eq a) => a

pattern GL_SAMPLE_COVERAGE :: (Num a, Eq a) => a

pattern GL_SAMPLE_COVERAGE_INVERT :: (Num a, Eq a) => a

pattern GL_SAMPLE_COVERAGE_VALUE :: (Num a, Eq a) => a

pattern GL_SOURCE0_ALPHA :: (Num a, Eq a) => a

pattern GL_SOURCE0_RGB :: (Num a, Eq a) => a

pattern GL_SOURCE1_ALPHA :: (Num a, Eq a) => a

pattern GL_SOURCE1_RGB :: (Num a, Eq a) => a

pattern GL_SOURCE2_ALPHA :: (Num a, Eq a) => a

pattern GL_SOURCE2_RGB :: (Num a, Eq a) => a

pattern GL_SUBTRACT :: (Num a, Eq a) => a

pattern GL_TEXTURE0 :: (Num a, Eq a) => a

pattern GL_TEXTURE1 :: (Num a, Eq a) => a

pattern GL_TEXTURE10 :: (Num a, Eq a) => a

pattern GL_TEXTURE11 :: (Num a, Eq a) => a

pattern GL_TEXTURE12 :: (Num a, Eq a) => a

pattern GL_TEXTURE13 :: (Num a, Eq a) => a

pattern GL_TEXTURE14 :: (Num a, Eq a) => a

pattern GL_TEXTURE15 :: (Num a, Eq a) => a

pattern GL_TEXTURE16 :: (Num a, Eq a) => a

pattern GL_TEXTURE17 :: (Num a, Eq a) => a

pattern GL_TEXTURE18 :: (Num a, Eq a) => a

pattern GL_TEXTURE19 :: (Num a, Eq a) => a

pattern GL_TEXTURE2 :: (Num a, Eq a) => a

pattern GL_TEXTURE20 :: (Num a, Eq a) => a

pattern GL_TEXTURE21 :: (Num a, Eq a) => a

pattern GL_TEXTURE22 :: (Num a, Eq a) => a

pattern GL_TEXTURE23 :: (Num a, Eq a) => a

pattern GL_TEXTURE24 :: (Num a, Eq a) => a

pattern GL_TEXTURE25 :: (Num a, Eq a) => a

pattern GL_TEXTURE26 :: (Num a, Eq a) => a

pattern GL_TEXTURE27 :: (Num a, Eq a) => a

pattern GL_TEXTURE28 :: (Num a, Eq a) => a

pattern GL_TEXTURE29 :: (Num a, Eq a) => a

pattern GL_TEXTURE3 :: (Num a, Eq a) => a

pattern GL_TEXTURE30 :: (Num a, Eq a) => a

pattern GL_TEXTURE31 :: (Num a, Eq a) => a

pattern GL_TEXTURE4 :: (Num a, Eq a) => a

pattern GL_TEXTURE5 :: (Num a, Eq a) => a

pattern GL_TEXTURE6 :: (Num a, Eq a) => a

pattern GL_TEXTURE7 :: (Num a, Eq a) => a

pattern GL_TEXTURE8 :: (Num a, Eq a) => a

pattern GL_TEXTURE9 :: (Num a, Eq a) => a

pattern GL_TEXTURE_BINDING_CUBE_MAP :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_IMAGE_SIZE :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSION_HINT :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_NEGATIVE_X :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_NEGATIVE_Y :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_NEGATIVE_Z :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_POSITIVE_X :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_POSITIVE_Y :: (Num a, Eq a) => a

pattern GL_TEXTURE_CUBE_MAP_POSITIVE_Z :: (Num a, Eq a) => a

pattern GL_TRANSPOSE_COLOR_MATRIX :: (Num a, Eq a) => a

pattern GL_TRANSPOSE_MODELVIEW_MATRIX :: (Num a, Eq a) => a

pattern GL_TRANSPOSE_PROJECTION_MATRIX :: (Num a, Eq a) => a

pattern GL_TRANSPOSE_TEXTURE_MATRIX :: (Num a, Eq a) => a