Graphics.GL.Raw.Extension.ARB.TextureCompression

Extension Support

gl_ARB_texture_compression :: Bool

Checks that the GL_ARB_texture_compression extension is available.

GL_ARB_texture_compression

glCompressedTexImage1DARB :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage1DARB target level internalformat width border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexImage1D.

glCompressedTexImage2DARB :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage2DARB target level internalformat width height border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexImage2D.

glCompressedTexImage3DARB :: MonadIO m => GLenum -> GLint -> GLenum -> GLsizei -> GLsizei -> GLsizei -> GLint -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexImage3DARB target level internalformat width height depth border imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter internalformat is a PixelInternalFormat.

The parameter border is a CheckedInt32.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexImage3D.

glCompressedTexSubImage1DARB :: MonadIO m => GLenum -> GLint -> GLint -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage1DARB target level xoffset width format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexSubImage1D.

glCompressedTexSubImage2DARB :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage2DARB target level xoffset yoffset width height format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexSubImage2D.

glCompressedTexSubImage3DARB :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> GLsizei -> GLenum -> GLsizei -> Ptr () -> m ()

Usage: glCompressedTexSubImage3DARB target level xoffset yoffset zoffset width height depth format imageSize data

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter xoffset is a CheckedInt32.

The parameter yoffset is a CheckedInt32.

The parameter zoffset is a CheckedInt32.

The parameter format is a PixelFormat, one of: GL_ABGR_EXT, GL_ALPHA, GL_BLUE, GL_CMYKA_EXT, GL_CMYK_EXT, GL_COLOR_INDEX, GL_DEPTH_COMPONENT, GL_GREEN, GL_LUMINANCE, GL_LUMINANCE_ALPHA, GL_RED, GL_RED_EXT, GL_RGB, GL_RGBA, GL_STENCIL_INDEX, GL_UNSIGNED_INT, GL_UNSIGNED_SHORT, GL_YCRCB_422_SGIX, GL_YCRCB_444_SGIX.

The parameter data is a CompressedTextureARB.

The length of data should be imageSize.

This command is an alias for glCompressedTexSubImage3D.

glGetCompressedTexImageARB :: MonadIO m => GLenum -> GLint -> Ptr () -> m ()

Usage: glGetCompressedTexImageARB target level img

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS.

The parameter level is a CheckedInt32.

The parameter img is a CompressedTextureARB.

The length of img should be COMPSIZE(target,level).

This command is an alias for glGetCompressedTexImage.

pattern GL_COMPRESSED_ALPHA_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_INTENSITY_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_LUMINANCE_ALPHA_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_LUMINANCE_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGBA_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_RGB_ARB :: (Num a, Eq a) => a

pattern GL_COMPRESSED_TEXTURE_FORMATS_ARB :: (Num a, Eq a) => a

pattern GL_NUM_COMPRESSED_TEXTURE_FORMATS_ARB :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_ARB :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSED_IMAGE_SIZE_ARB :: (Num a, Eq a) => a

pattern GL_TEXTURE_COMPRESSION_HINT_ARB :: (Num a, Eq a) => a