Graphics.GL.Raw.Extension.EXT.CopyTexture

Extension Support

gl_EXT_copy_texture :: Bool

GL_EXT_copy_texture

glCopyTexImage1DEXT :: MonadIO m => GLenum -> GLint -> GLenum -> GLint -> GLint -> GLsizei -> GLint -> m ()

Usage: glCopyTexImage1DEXT target level internalformat x y width border

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS

The parameter level is a CheckedInt32

The parameter internalformat is a PixelInternalFormat

The parameter x is a WinCoord

The parameter y is a WinCoord

The parameter border is a CheckedInt32

This command is an alias for glCopyTexImage1D

glCopyTexImage2DEXT :: MonadIO m => GLenum -> GLint -> GLenum -> GLint -> GLint -> GLsizei -> GLsizei -> GLint -> m ()

Usage: glCopyTexImage2DEXT target level internalformat x y width height border

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS

The parameter level is a CheckedInt32

The parameter internalformat is a PixelInternalFormat

The parameter x is a WinCoord

The parameter y is a WinCoord

The parameter border is a CheckedInt32

This command is an alias for glCopyTexImage2D

glCopyTexSubImage1DEXT :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLsizei -> m ()

Usage: glCopyTexSubImage1DEXT target level xoffset x y width

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS

The parameter level is a CheckedInt32

The parameter xoffset is a CheckedInt32

The parameter x is a WinCoord

The parameter y is a WinCoord

This command is an alias for glCopyTexSubImage1D

glCopyTexSubImage2DEXT :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage2DEXT target level xoffset yoffset x y width height

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS

The parameter level is a CheckedInt32

The parameter xoffset is a CheckedInt32

The parameter yoffset is a CheckedInt32

The parameter x is a WinCoord

The parameter y is a WinCoord

This command is an alias for glCopyTexSubImage2D

glCopyTexSubImage3DEXT :: MonadIO m => GLenum -> GLint -> GLint -> GLint -> GLint -> GLint -> GLint -> GLsizei -> GLsizei -> m ()

Usage: glCopyTexSubImage3DEXT target level xoffset yoffset zoffset x y width height

The parameter target is a TextureTarget, one of: GL_DETAIL_TEXTURE_2D_SGIS, GL_PROXY_TEXTURE_1D, GL_PROXY_TEXTURE_1D_EXT, GL_PROXY_TEXTURE_2D, GL_PROXY_TEXTURE_2D_EXT, GL_PROXY_TEXTURE_3D, GL_PROXY_TEXTURE_3D_EXT, GL_PROXY_TEXTURE_4D_SGIS, GL_TEXTURE_1D, GL_TEXTURE_2D, GL_TEXTURE_3D, GL_TEXTURE_3D_EXT, GL_TEXTURE_3D_OES, GL_TEXTURE_4D_SGIS, GL_TEXTURE_BASE_LEVEL, GL_TEXTURE_BASE_LEVEL_SGIS, GL_TEXTURE_MAX_LEVEL, GL_TEXTURE_MAX_LEVEL_SGIS, GL_TEXTURE_MAX_LOD, GL_TEXTURE_MAX_LOD_SGIS, GL_TEXTURE_MIN_LOD, GL_TEXTURE_MIN_LOD_SGIS

The parameter level is a CheckedInt32

The parameter xoffset is a CheckedInt32

The parameter yoffset is a CheckedInt32

The parameter zoffset is a CheckedInt32

The parameter x is a WinCoord

The parameter y is a WinCoord

This command is an alias for glCopyTexSubImage3D